% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccumulatR.R
\name{AccumulatR_model}
\alias{AccumulatR_model}
\title{Use an AccumulatR model in EMC2}
\usage{
AccumulatR_model(model)
}
\arguments{
\item{model}{A finalized AccumulatR model.}
}
\value{
An EMC2 model function suitable for \code{design()}.
}
\description{
Use an AccumulatR model in EMC2
}
