% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{adaptive_pars}
\alias{adaptive_pars}
\title{Get adaptive parameter names}
\usage{
adaptive_pars(adaptive)
}
\arguments{
\item{adaptive}{A list containing a dyntype specification}
}
\value{
List of names generated from an adaptive "dyntype" specification
}
\description{
Get adaptive parameter names
}
