% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{add_accumulators}
\alias{add_accumulators}
\title{add_accumulators()}
\usage{
add_accumulators(data, matchfun = NULL, simulate = FALSE, type = "RACE")
}
\arguments{
\item{type}{}
}
\description{
Augments data for use in race model likelihood calculation or simulation.
Must have an R = response factor (levels give number of accumulators).
Must have a "subjects" factor and optionally a "trials" indicator column with
unique values for each subject and trial (or trial within cell) (typically
integer), then any number of factors. For the default race model type
replicates data by number response levels, and adds accumulator factor lR.
If matchfun is supplied adds a match factor (lM) scored by matchfun.
For likelihood should have an rt column, R should have values and a "winner"
logical column (latent and observed response agree) is added.
For simulate R need not have values and an rt column is added.
If not type RACE (e.g., DDM type) doesn't replicate data and adds dummy lR,
lM and winner columns.
}
