% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{add_constants}
\alias{add_constants}
\title{Augments parameter matrix or vector p with constant parameters (also used in data)}
\usage{
add_constants(p, constants)
}
\arguments{
\item{p}{either a matrix or vector of parameters}

\item{constants}{a named vector of constants}
}
\value{
a matrix or vector, depending on input, with the varying parameters and constants combined.
}
\description{
Augments parameter matrix or vector p with constant parameters (also used in data)
}
