% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{add_ICs_MLL}
\alias{add_ICs_MLL}
\title{Add information criteria to emc object}
\usage{
add_ICs_MLL(
  emc,
  stage = "sample",
  filter = NULL,
  use_best_fit = TRUE,
  BayesFactor = TRUE,
  cores_for_props = 4,
  cores_per_prop = 1,
  ...
)
}
\arguments{
\item{emc}{List of samples objects}

\item{stage}{A string. Specifies which stage the samples are to be taken from \code{"preburn"}, \code{"burn"}, \code{"adapt"}, or \code{"sample"}}

\item{filter}{An integer or vector. If it's an integer, iterations up until the value set by \code{filter} will be excluded.
If a vector is supplied, only the iterations in the vector will be considered.}

\item{use_best_fit}{Boolean, defaults to \code{TRUE}, uses the minimal or mean likelihood (whichever is better) in the
calculation, otherwise always uses the mean likelihood.}

\item{BayesFactor}{Boolean, defaults to \code{TRUE}. Include marginal likelihoods as estimated using WARP-III bridge sampling.
Usually takes a minute per model added to calculate}

\item{cores_for_props}{Integer, how many cores to use for the Bayes factor calculation, here 4 is the default for the 4 different proposal densities to evaluate, only 1, 2 and 4 are sensible.}

\item{cores_per_prop}{Integer, how many cores to use for the Bayes factor calculation if you have more than 4 cores available. Cores used will be cores_for_props * cores_per_prop. Best to prioritize cores_for_props being 4 or 2}
}
\value{
An \code{emc} object with new attributes 'ICs' and 'MLL':
}
\description{
Adds DIC, BPIC, and optionally MLL values as attributes to an emc object. Can be useful to offload
computational burden.
}
