% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{add_pw_ll}
\alias{add_pw_ll}
\title{Add Pointwise Log-Likelihoods to an emc Object}
\usage{
add_pw_ll(emc, cores_for_chains = 1)
}
\arguments{
\item{emc}{An emc object.}

\item{cores_for_chains}{Integer. Number of cores to use for parallel calculation across chains.}
}
\value{
The same emc object with \code{pw_ll} added to each chain's samples.
}
\description{
Calculates the pointwise log-likelihoods for all samples in the emc object
and stores them in the \code{samples} list of each chain. This allows
for faster subsequent calls to \code{compare()} or \code{extract_pw_ll()}.
}
