% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_funcs.R
\name{auto_thin.emc}
\alias{auto_thin.emc}
\alias{auto_thin}
\title{Automatically thin an emc object}
\usage{
\method{auto_thin}{emc}(emc, stage = "sample", selection = c("alpha", "mu"))

auto_thin(emc, stage = "sample", selection = c("alpha", "mu"), ...)
}
\arguments{
\item{emc}{an emc object.}

\item{stage}{A character string. Indicates from which sampling stage(s) to take the samples from (i.e. \code{preburn}, \code{burn}, \code{adapt}, \code{sample})}

\item{selection}{Which parameter types (i.e. 'alpha' or 'mu' to consider when determining the effective sample size)}

\item{...}{additional optional arguments}
}
\description{
Uses the effective sample size of \code{selection} to determine how much to optimally thin an emc object
}
