% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MTR.R
\name{BE2LBA}
\alias{BE2LBA}
\title{Balance of Evidence 2 Threshold LBA model}
\usage{
BE2LBA()
}
\value{
A model list with all the necessary functions to sample
}
\description{
All thresholds greater than A, DT1 <- A+DT1, b <- DT1 + B
}
