% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LBA.R
\name{BE2lbaBP}
\alias{BE2lbaBP}
\title{Balance of Evidence 2 Threshold LBA model}
\usage{
BE2lbaBP()
}
\value{
A model list with all the necessary functions to sample
}
\description{
Intermediate thresholds proportional to b, can be less than A
}
