% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MTR.R
\name{BE2PLBA}
\alias{BE2PLBA}
\title{Balance of Evidence 2 Threshold LBA model}
\usage{
BE2PLBA()
}
\value{
A model list with all the necessary functions to sample
}
\description{
Intermediate thresholds and gb proportional to B
}
