% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LBA.R
\name{BE3lbaB}
\alias{BE3lbaB}
\title{Balance of Evidence 3 Threshold LBA model}
\usage{
BE3lbaB()
}
\value{
A model list with all the necessary functions to sample
}
\description{
All thresholds greater than A, DT1 <- A+DT1, DT2 <- DT1 + DT2, b <- DT2 + B
}
