% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MTR.R
\name{BE3PLBA}
\alias{BE3PLBA}
\title{Balance of Evidence 3 Threshold LBA model}
\usage{
BE3PLBA()
}
\value{
A model list with all the necessary functions to sample
}
\description{
Intermediate thresholds proportional to b, can be less than A
}
