% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MTR.R
\name{BE4LBA}
\alias{BE4LBA}
\title{Balance of Evidence 4 Threshold LBA model}
\usage{
BE4LBA()
}
\value{
A model list with all the necessary functions to sample
}
\description{
All thresholds greater than A, DT1 <- A+DT1, DT2 <- DT1 + DT2, DT3 <- DT2 + DT3, b <- DT3 + B
}
