% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LBA.R
\name{BE4lbaBP}
\alias{BE4lbaBP}
\title{Balance of Evidence 4 Threshold LBA model}
\usage{
BE4lbaBP()
}
\value{
A model list with all the necessary functions to sample
}
\description{
Intermediate thresholds proportional to b, can be less than A
}
