% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MTR.R
\name{BE4PLBA}
\alias{BE4PLBA}
\title{Balance of Evidence 4 Threshold LBA model}
\usage{
BE4PLBA()
}
\value{
A model list with all the necessary functions to sample
}
\description{
Intermediate thresholds proportional to b, can be less than A
}
