% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{calc_ll_pw}
\alias{calc_ll_pw}
\title{Calculate Pointwise Log-Likelihoods}
\usage{
calc_ll_pw(proposals, dadm, model, r_cores = 1)
}
\arguments{
\item{proposals}{A matrix of parameter proposals with dimensions \verb{n_iter x n_pars}.}

\item{dadm}{A Data Augmented Design Matrix (DADM).}

\item{model}{A model function or object.}

\item{r_cores}{Number of cores to use (for non-C models).}
}
\value{
A matrix of pointwise log-likelihoods with dimensions \verb{n_iter x n_trials}.
}
\description{
Calculate Pointwise Log-Likelihoods
}
