% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_standard.R
\name{calculate_implied_means}
\alias{calculate_implied_means}
\title{Calculate implied means}
\usage{
calculate_implied_means(mean_designs, beta_params, n_pars)
}
\arguments{
\item{mean_designs}{List of mean design matrices}

\item{beta_params}{Vector of beta parameters (all parameters including regressors)}

\item{n_pars}{Number of parameters}
}
\value{
Vector of implied means
}
\description{
Calculates the implied means for parameters using mean design matrices
}
