% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_standard.R
\name{calculate_mean_design}
\alias{calculate_mean_design}
\title{Calculate mean design matrices}
\usage{
calculate_mean_design(group_designs, n_pars)
}
\arguments{
\item{group_designs}{List of design matrices for each parameter}

\item{n_pars}{Number of parameters}
}
\value{
List of mean design matrices
}
\description{
Calculates the mean design matrix for each parameter across subjects
}
