% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_CDM.R
\name{CDM}
\alias{CDM}
\title{The Circular Diffusion Model (CDM)}
\usage{
CDM()
}
\value{
A model list with all the necessary functions for EMC2 to sample.
}
\description{
Model file to estimate the Circular Diffusion Model (CDM) in EMC2.
}
\details{
Parameters and defaults are on the transformed scale.
All parameters not specified in the \code{formula} of \code{design()} are
assumed constant with the defaults in \code{p_types}.\tabular{lllll}{
   Parameter \tab Transform \tab Natural scale \tab Default \tab Interpretation \cr
   v \tab log \tab [0, Inf) \tab log(1) \tab Drift magnitude \cr
   theta \tab pnorm \tab [0, 1) \tab qnorm(.5) \tab Drift direction, mapped to -pi, pi \cr
   a \tab log \tab [0, Inf) \tab log(1) \tab Boundary radius \cr
   t0 \tab log \tab [0, Inf) \tab log(0) \tab Non-decision time \cr
   s \tab log \tab [0, Inf) \tab log(1) \tab Diffusion scale \cr
   sv \tab log \tab [0, Inf) \tab log(0) \tab Drift SD across trials \cr
}
}
