% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{chain_n}
\alias{chain_n}
\title{chain_n()}
\usage{
chain_n(samplers)
}
\arguments{
\item{samplers}{A list, the output of \code{run_emc()}.}
}
\value{
A matrix
}
\description{
Returns a matrix with the number of samples per chain for each stage that is present
in the sampling object (i.e., \code{preburn}, \code{burn}, \code{adapt},
\code{sample}). The number of rows of the matrix reflects the number of chains
and the number of columns the number of sampling stages.
}
\examples{
\dontrun{
chain_n(samplers)
}
}
