% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{check_reparameterisations}
\alias{check_reparameterisations}
\title{Check and update formula list for reparameterisation parameters}
\usage{
check_reparameterisations(reparameterisations, formula, constants = NULL)
}
\arguments{
\item{reparameterisations}{Named list, e.g.
list(v = c(new_par1=1, new_par2=0.5), B = c(new_par1=1, new_par2=-0.5))}

\item{formula}{List of formulas (LHS = model parameter names)}

\item{constants}{Named vector/list of constants, or NULL}
}
\value{
Updated formula list with reparameterisationsd source parameters added as
intercept-only formulas where missing
}
\description{
Check and update formula list for reparameterisation parameters
}
