% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{check_run_new}
\alias{check_run_new}
\title{Convergence checks for an EMC2 samplers object}
\usage{
check_run_new(
  samplers,
  selection = c("mu", "sigma2", "alpha"),
  digits = 3,
  plot_worst = TRUE,
  ...
)
}
\description{
Runs a series of convergence checks, prints statistics to the console, and
saves traceplots to a PDF file.
}
\details{
Note that the \code{Rhat} is calculated by doubling the number of chains by
first splitting chains into first and second half, so it also a test of
stationarity.

Efficiency of sampling is (optionally) indicated by integrated
autocorrelation time (from the \code{LaplacesDemon} R package) and by the effective
number of samples (from the \code{coda} R package).
}
