% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{contr.anova}
\alias{contr.anova}
\title{contr.anova()}
\usage{
contr.anova(n)
}
\arguments{
\item{n}{an integer. the number of items for which to create the contrast}
}
\value{
a contrast matrix.
}
\description{
orthogonal helmert contrast scaled to estimate differences between conditions. Use in make_design.
}
