% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{contr.decreasing}
\alias{contr.decreasing}
\title{Contrast to enforce decreasing estimates}
\usage{
contr.decreasing(n, levels = NULL)
}
\arguments{
\item{n}{an integer. The number of items for which to create the contrast.}

\item{levels}{Character vector. the factor levels which will be the colnames of the returning matrix.}
}
\value{
a contrast matrix.
}
\description{
Contrast to enforce decreasing estimates
}
