% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{contr.increasing}
\alias{contr.increasing}
\title{contr.increasing()}
\usage{
contr.increasing(n, levels = NULL)
}
\arguments{
\item{levels}{}
}
\description{
special contrast for SDT threshold factor
first = intercept, cumsum other (positive) levels to force non-decreasing
}
