% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{cut_factors}
\alias{cut_factors}
\title{Cut Factors Based on Credible Loadings}
\usage{
cut_factors(emc, CI = 95)
}
\arguments{
\item{emc}{An 'emc' object containing factor analysis results}

\item{CI}{Numeric. Confidence interval percentage (default is 95)}
}
\value{
An 'emc' object with factors that don't meet the credibility criterion removed
}
\description{
This function removes factors that do not have more than one credible loading
based on the specified confidence interval.
}
