% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_DDM.R
\name{DDMGNGnoC}
\alias{DDMGNGnoC}
\title{The GNG Diffusion Decision Model}
\usage{
DDMGNGnoC()
}
\value{
A model list with all the necessary functions to sample
}
\description{
The GNG Diffusion Decision Model
}
