% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_DDM.R
\name{ddmTZDt0natural}
\alias{ddmTZDt0natural}
\title{Diffusion decision model with t0 on the natural scale}
\usage{
ddmTZDt0natural()
}
\value{
A model list with all the necessary functions to sample
}
\description{
Diffusion decision model with t0 on the natural scale
}
