% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMC2-package.R
\docType{package}
\name{EMC2-package}
\alias{EMC2}
\alias{EMC2-package}
\title{EMC2: Bayesian cognitive models of choice}
\description{
Fit Bayesian (hierarchical) cognitive models using a linear modeling language interface. The DDM, LBA, RDM, LNR and SDT model are supported, and the user can also specify their own likelihood function. Additionally, the user can choose for non-hierarchical estimation, as well as a diagonal, blocked or full multivariate normal group-level distribution to test individual differences. Prior specification is facilitated through methods that visualize the (implied) prior. A wide range of plotting functions assist in assessing model convergence and posterior inference. Models can be easily evaluated using functions that plot posterior predictions or using relative model comparison metrics such as information criteria or Bayes factors. References: Stevenson et al. (2024) <doi.org/10.31234/osf.io/2e4dq>.
}
\author{
\strong{Maintainer}: Niek Stevenson \email{niek.stevenson@gmail.com}

Authors:
\itemize{
  \item Michelle Donzallaz
  \item Andrew Heathcote
}

}
\keyword{internal}
