% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{emc_algorithm_libraries}
\alias{emc_algorithm_libraries}
\title{Given a vector of algorithms, return all the libraries needed by those algorithms}
\usage{
emc_algorithm_libraries(algorithms, uninstalled_only = TRUE)
}
\arguments{
\item{algorithms}{Algorithms to list libraries for}

\item{uninstalled_only}{If TRUE only uninstalled libraries are returned.}
}
\value{
Vector of library names
}
\description{
Given a vector of algorithms, return all the libraries needed by those algorithms
}
