% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{emc_algorithms}
\alias{emc_algorithms}
\title{List emc algorithms}
\usage{
emc_algorithms(uninstalled = FALSE)
}
\arguments{
\item{uninstalled}{If TRUE list include uninstalled algorithms}
}
\value{
A dataframe listing available
}
\description{
\code{emc_algorithms} lists all the algorithms available to emc.
}
\details{
As well as the algorithm it lists a set of tags useful for filtering the
list of algorithms.
}
