% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recording.R
\name{emc_record}
\alias{emc_record}
\title{Records a function}
\usage{
emc_record(f, label = NULL, verbose = TRUE, cache_path = NULL)
}
\value{
A tibble that contains the result, error, execution times and the replay log.
}
\description{
Executes an function  recording all messages, warnings and errors along the way.
}
