% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{es_summary}
\alias{es_summary}
\title{es_summary}
\usage{
es_summary(samplers, no_print = FALSE, return_min = FALSE, digits = 2)
}
\arguments{
\item{samplers}{Samples object with multiple chains}

\item{no_print}{Boolean for printing}

\item{return_min}{return is min(es) for hierarchical}

\item{digits}{Integer, number of digits for printing}
}
\description{
Summarizes effective size statistics for a samplers object, invisibly
returning as a list
}
