% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{extract_pw_ll}
\alias{extract_pw_ll}
\title{Extract Pointwise Log-Likelihoods}
\usage{
extract_pw_ll(emc, stage = "sample", filter = 0, FUN = mean)
}
\arguments{
\item{emc}{An emc object.}

\item{stage}{A string. Specifies which stage the samples are to be taken from \code{"preburn"}, \code{"burn"}, \code{"adapt"}, or \code{"sample"}.}

\item{filter}{An integer or vector. iterations to exclude.}

\item{FUN}{A function to aggregate across iterations (e.g., \code{mean}, \code{median}, or \code{NULL} to return all iterations).}
}
\value{
A data frame (concatenated DADMs) with an additional \code{ll} column (or multiple columns if \code{FUN=NULL}).
}
\description{
Extracts the pointwise log-likelihoods for each trial in the data and
appends them to the Data Augmented Design Matrix (DADM).
}
