% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forstmann}
\alias{forstmann}
\title{Forstmann et al.'s data}
\format{
A data frame with 15818 rows and 5 variables:
\describe{
\item{E}{Factor with 3 levels for Speed, Accuracy and
Neutral}
\item{R}{Factor with 2 levels for Left and Right responses}
\item{S}{Factor with 2 levels for Left and Right trials}
\item{rt}{reaction time for each trial as a double}
\item{subjects}{integer ID for each subject}
}
}
\source{
\url{https://www.pnas.org/doi/10.1073/pnas.0805903105}
}
\usage{
forstmann
}
\description{
A dataset containing the speed or accuracy manipulation for a Random Dot
Motion experiment.
}
\details{
Details on the dataset can be found in the following paper:

\strong{Striatum and pre-SMA facilitate decision-making under time pressure}

Birte U. Forstmann, Gilles Dutilh, Scott Brown, Jane Neumann,
D. Yves von Cramon, K. Richard Ridderinkhof, Eric-Jan Wagenmakers.

\emph{Proceedings of the National Academy of Sciences Nov 2008, 105 (45)
17538-17542; DOI: 10.1073/pnas.0805903105}
}
\keyword{datasets}
