% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{gd}
\alias{gd}
\title{gd}
\usage{
gd(samplers, no_print = FALSE, return_max = FALSE, digits = 2)
}
\arguments{
\item{samplers}{Samples object with multiple chains}

\item{no_print}{Boolean for printing}

\item{return_max}{return is max(gd) for hierarchical}

\item{digits}{Integer, number of digits for printing}
}
\description{
Summarizes gelman_diag statistics for a samplers object, invisibly returning
a list of two lists containing univarite (psrf) and multivariate (mpsrf)
statistics.
}
