% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{gd_pmwg}
\alias{gd_pmwg}
\title{Title}
\usage{
gd_pmwg(
  pmwg_mcmc,
  return_summary = FALSE,
  print_summary = TRUE,
  digits_print = 2,
  sort_print = TRUE,
  autoburnin = FALSE,
  transform = TRUE,
  selection = "alpha",
  filter = "sample",
  thin = 1,
  subfilter = NULL,
  mapped = FALSE
)
}
\arguments{
\item{mapped}{}
}
\description{
Title
}
