% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_RDM.R
\name{generateRDMdynamic}
\alias{generateRDMdynamic}
\title{RL-RD with RDM as model}
\usage{
generateRDMdynamic(
  p_types,
  constantsFixed = NULL,
  probit_scale = c("alpha1", "alpha2", "alpha3", "q01", "q02", "q03"),
  logit_scale = c("v", "B", "A", "t0", "s")
)
}
\description{
RL-RD with RDM as model
}
