% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{get_adaptive}
\alias{get_adaptive}
\title{Get adaptive parameter, or underlying non-linear components}
\usage{
get_adaptive(
  pname,
  p_vector,
  data,
  design,
  return_dadm = TRUE,
  return_p = TRUE,
  compress = FALSE
)
}
\arguments{
\item{pname}{parameter to provide results for}

\item{p_vector}{parameter vector}

\item{data}{data frame providing covariate(s)}

\item{design}{design with dynamic RL component}

\item{return_dadm}{return full dadm (accumulators) otherwise just trials}

\item{return_p}{return dynamic parameter, default FALSE, return non-linear}

\item{compress}{compress dadm? Default FALSE appropriate for RL models}
}
\value{
dadm/data q values, prediction errors etc. (etc. depends on RL model dyntype)
}
\description{
Non-linear components are q values and prediction errors for RL models and
kernel for trend models
}
