% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_sampling.R
\name{get_BayesFactor}
\alias{get_BayesFactor}
\title{get_BayesFactor}
\usage{
get_BayesFactor(MLL1, MLL2)
}
\arguments{
\item{MLL1}{Numeric. Marginal likelihood of model 1. Obtained with \code{run_bridge_sampling}}

\item{MLL2}{Numeric. Marginal likelihood of model 1}
}
\value{
The BayesFactor for model 1 over model 2
}
\description{
returns the Bayes Factor for two models
}
