% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_funcs.R
\name{get_data.emc}
\alias{get_data.emc}
\alias{get_data}
\title{Get data}
\usage{
\method{get_data}{emc}(emc)

get_data(emc)
}
\arguments{
\item{emc}{an emc object}
}
\value{
A dataframe of the original data
}
\description{
Extracts data from an emc object
}
\details{
emc adds columns and rows to a dataframe in order to facilitate efficient likelihood calculations.
This function will return the data as provided originally.
}
\examples{
get_data(samples_LNR)
}
