% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R, R/s3_funcs.R
\name{get_design.emc.prior}
\alias{get_design.emc.prior}
\alias{get_design.emc}
\alias{get_design}
\title{Get Design}
\usage{
\method{get_design}{emc.prior}(x)

\method{get_design}{emc}(x)

get_design(x)
}
\arguments{
\item{x}{an \code{emc} or \code{emc.prior} object}
}
\value{
A design with class emc.design
}
\description{
Extracts design from an emc object
}
\examples{
get_design(samples_LNR)
}
