% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R, R/s3_funcs.R
\name{get_group_design.emc.prior}
\alias{get_group_design.emc.prior}
\alias{get_group_design.emc}
\alias{get_group_design}
\title{Get Group Design}
\usage{
\method{get_group_design}{emc.prior}(x)

\method{get_group_design}{emc}(x)

get_group_design(x)
}
\arguments{
\item{x}{an \code{emc} or \code{emc.prior} object}
}
\value{
A design with class emc.group_design
}
\description{
Extracts group design from an emc object
}
