% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_factor.R
\name{get_prior_factor}
\alias{get_prior_factor}
\title{Prior specification or prior sampling for factor estimation.}
\usage{
get_prior_factor(
  prior = NULL,
  n_pars = NULL,
  sample = TRUE,
  N = 1e+05,
  type = "mu",
  design = NULL,
  map = FALSE,
  n_factors = 5
)
}
\arguments{
\item{prior}{A named list containing the prior mean (theta_mu_mean) and
variance (theta_mu_var). Default prior created if NULL}

\item{n_pars}{Argument used by the sampler, best left NULL. In user case inferred from the design}

\item{sample}{Whether to sample from the prior. Default is TRUE. If not returns a prior list}

\item{N}{How many samples to draw from the prior, default 1e5}

\item{type}{character, options: "mu", "variance", "covariance" "full_var"}

\item{design}{The design obtained from \code{make_design}, required when map = TRUE}

\item{map}{Boolean, default TRUE reverses malformation used by model to make
sampled parameters unbounded}

\item{n_factors}{integer. The number of factors.}
}
\value{
A list with a single entry of type of samples from the prior (if sample = TRUE) or else a prior object
}
\description{
Prior specification or prior sampling for factor estimation.
}
