% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{get_stop_criteria}
\alias{get_stop_criteria}
\title{Set or return the stop criteria associated with a stage}
\usage{
get_stop_criteria(stage = "sample", stop_criteria = NULL, type = "standard")
}
\arguments{
\item{stage}{A string. Indicates which stage to get the stop_criteria for either \code{preburn}, \code{burn}, \code{adapt} or \code{sample}.}

\item{stop_criteria}{A list with stopping criteria. If \code{NULL} will fill defaults. If already partially specified will fill the rest with defaults.
Also see the details and examples section of \code{run_emc}.}

\item{type}{A string indicating whether to run a \code{standard} group-level, \code{blocked}, \code{diagonal}, \code{factor}, or \code{single} (i.e., non-hierarchical) model.
The stopping criteria vary between hierarchical and non-hierarchical models (\code{type = single})}
}
\value{
A list with stopping criteria associated with the stage
}
\description{
Set or return the stop criteria associated with a stage
}
\examples{
# To see the stop criteria for e.g. the adapt phase
get_stop_criteria("adapt")
# Or for all stages:
stages <- c("preburn", "burn", "adapt", "sample")
stop_criteria <- lapply(stages, get_stop_criteria)
names(stop_criteria) <- stages

}
