% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{graphical_model}
\alias{graphical_model}
\title{Graphical Model}
\usage{
graphical_model(emc)
}
\arguments{
\item{emc}{an emc object of type \code{standard}, \code{blocked}, or \code{diagonal}}
}
\value{
Invisibly returns a DiagrammeR \code{grViz} graph object
}
\description{
Draws a probabilistic graphical model (PGM) using circles, squares, arrows,
and plates for the standard EMC hierarchical model.
}
