% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_design.R
\name{has_random}
\alias{has_random}
\title{Check if group design has random effects}
\usage{
has_random(group_design)
}
\arguments{
\item{group_design}{A group design object}
}
\value{
Boolean
}
\description{
Check if group design has random effects
}
