% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{hyper_prediction}
\alias{hyper_prediction}
\title{hyper prediction}
\usage{
hyper_prediction(
  pmwgs,
  N = 1,
  n_trials = 100,
  return_pars = FALSE,
  use_prior = FALSE,
  max_trys = 100
)
}
\arguments{
\item{pmwgs}{and emc object}

\item{N}{number of participants}

\item{n_trials}{number of trials per cell}

\item{return_pars}{return pars instead of simulated data}

\item{use_prior}{logical, default FALSE, if TRUE sample from prior}

\item{max_trys}{mmaximum trys to get valid parameters (in lots of N)}
}
\value{
simulated data or a parameter matrix
}
\description{
creates data sets by sampling from the posterior of an emc object
}
