% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{iat}
\alias{iat}
\title{iat}
\usage{
iat(samplers, no_print = FALSE, return_max = FALSE, digits = 2)
}
\arguments{
\item{samplers}{Samples object with multiple chains}

\item{no_print}{Boolean for printing}

\item{return_max}{return is max(iat) for hierarchical}

\item{digits}{Integer, number of digits for printing}
}
\description{
Summarizes integrated autocorrelation time statistics for a samplers object,
invisibly returning as a list
}
