% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{IC}
\alias{IC}
\title{Title}
\usage{
IC(
  samplers,
  filter = "sample",
  subfilter = 0,
  use_best_fit = TRUE,
  print_summary = TRUE,
  digits = 0,
  subject = NULL
)
}
\arguments{
\item{subject}{}
}
\description{
Title
}
