% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IS2.R
\name{IS2}
\alias{IS2}
\title{Title}
\usage{
IS2(
  samples,
  filter = "sample",
  subfilter = 0,
  IS_samples = 1000,
  stepsize_particles = 500,
  max_particles = 5000,
  n_cores = 1,
  df = 5
)
}
\arguments{
\item{df}{}
}
\description{
Title
}
