% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LBA.R
\name{LBAIO}
\alias{LBAIO}
\title{LBAIO
LBA model accommodating missing values (truncation and censoring) and
assuming unbounded rates (i.e., allows intrinsic omissions)}
\usage{
LBAIO()
}
\description{
LBAIO
LBA model accommodating missing values (truncation and censoring) and
assuming unbounded rates (i.e., allows intrinsic omissions)
}
