% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LNR.R
\name{LNRcorr}
\alias{LNRcorr}
\title{Correlated Log-Normal Race Model}
\usage{
LNRcorr()
}
\description{
Convenience constructor for the correlated form of the log-normal race.
}
\details{
The two participating accumulator finishing times are coupled by a Gaussian
copula with natural-scale correlation parameter \code{rho}. All other
accumulators remain independent. This is equivalent to
\code{LNR(correlated = TRUE)}.
}
\value{
A model list compatible with \code{design()}.
}
\seealso{\code{\link{LNR}}}
