% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{log_likelihood_joint}
\alias{log_likelihood_joint}
\title{log_likelihood_joint}
\usage{
log_likelihood_joint(proposals, dadms, component = NULL)
}
\arguments{
\item{component}{}
}
\value{
summed log-likelihood
}
\description{
log_likelihood_joint
}
