% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fMRI.R
\name{make_fmri_design_matrix}
\alias{make_fmri_design_matrix}
\title{Function to create fMRI design matrix from a few basics}
\usage{
make_fmri_design_matrix(
  frame_times,
  events = NULL,
  hrf_model = "glover",
  drift_model = "cosine",
  high_pass = 0.01,
  drift_order = 1,
  fir_delays = c(0),
  add_regs = NULL,
  add_reg_names = NULL,
  min_onset = -24,
  oversampling = 50,
  add_intercept = TRUE
)
}
\description{
Function to create fMRI design matrix from a few basics
}
