% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{make_samplers}
\alias{make_samplers}
\title{\code{make_samplers()}}
\usage{
make_samplers(
  data,
  design,
  model = NULL,
  type = "standard",
  n_chains = 3,
  compress = TRUE,
  rt_resolution = 0.02,
  prior_list = NULL,
  grouped_pars = NULL,
  par_groups = NULL,
  n_factors = NULL,
  constraintMat = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame, or a list of data frames. Needs to have the variable \code{subjects} as participant identifier.}

\item{design}{A list with a pre-specified design, the output of \code{make_design()}.}

\item{model}{A model list. If none is supplied, the model specified in \code{make_design()} is used.}

\item{type}{A string indicating whether to run a \code{standard} group-level, \code{blocked}, \code{diagonal}, \code{factor}, or \code{single} (i.e., non-hierarchical) model.}

\item{n_chains}{An integer. Specifies the number of mcmc chains to be run (has to be more than 1 to compute \code{rhat}).}

\item{compress}{A Boolean, if \code{TRUE} (i.e., the default), the data is compressed to speed up likelihood calculations}

\item{rt_resolution}{A double. Used for compression, response times will be binned based on this resolution.}

\item{prior_list}{A named list containing the prior. Default prior created if NULL}

\item{grouped_pars}{An integer vector. Parameters on this location of the vector of parameters are treated as constant across sujects}

\item{par_groups}{A vector. Only to be specified with type blocked, e.g., \code{c(1,1,1,2,2)} means the covariances
of the first three and of the last two parameters are estimated as two separate blocks.}

\item{n_factors}{An integer. Only to be specified with type factor.}

\item{constraintMat}{A matrix of rows equal to the number of estimated parameters, and columns equal to the number of factors, only to be specified with type factor.
If null will use default settings as specified in Innes et al. 2022}

\item{...}{Additional, optional arguments}
}
\value{
a list of samplers
}
\description{
initializes the sampling process by combining the data, prior,
and model specification into a \code{samplers} object that is needed in \code{run_emc()}.
}
