% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{make_SEM_diagram}
\alias{make_SEM_diagram}
\title{Make SEM Diagram}
\usage{
make_SEM_diagram(
  emc,
  plot_values = TRUE,
  cred_only = FALSE,
  par_names = NULL,
  cut = NULL,
  ...
)
}
\arguments{
\item{emc}{an emc object}

\item{plot_values}{whether to plot the values or just the nodes/edges}

\item{cred_only}{whether to only plot credible values}

\item{par_names}{optional, if specified will overwrite the parameter names with
user-defined names}

\item{cut}{optional. A numeric value factor loadings smaller than this value will be omitted.}

\item{...}{optional additional arguments passed to render_graph from DiagrammeR}
}
\value{
Invisibly returns a DiagrammeR graph object
}
\description{
Make SEM Diagram
}
