% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mapped_par}
\alias{mapped_par}
\title{Maps a parameter vector through a design to show parameter values for each
design cell.}
\usage{
mapped_par(
  p_vector,
  design,
  model = NULL,
  digits = 3,
  remove_subjects = TRUE,
  Fcovariates = NULL
)
}
\arguments{
\item{p_vector}{The parameter vector (must have form of sampled_p_vector(design)}

\item{design}{The design (created by make_design)}

\item{model}{Optional model type (if not in design)}

\item{digits}{For rounding outputs}

\item{remove_subjects}{Default true removes subjects from output design}

\item{Fcovariates}{Any covariates in the design}
}
\value{
Matrix of factors and mapped parameter values
}
\description{
Maps a parameter vector through a design to show parameter values for each
design cell.
}
