% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_DDM.R
\name{MDDM}
\alias{MDDM}
\title{The DDM with response omissions}
\usage{
MDDM()
}
\value{
A list defining the cognitive model
}
\description{
The DDM with response omissions
}
\details{
See ?DDM for most details

pContaminant is a probit scaled parameter for non-process (contaminant) miss.
For background see :

Damaso, K. A. M., Castro, S. C., Todd, J., Strayer, D. L., Provost, A.,
Matzke, D., & Heathcote, A. (2021). A cognitive model of response omissions
in distraction paradigms. Memory & Cognition, 1–17.
https://doi.org/10.3758/s13421-021-01265-z

Note: This model does not have an Rcpp version.
}
