% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{merge_chains}
\alias{merge_chains}
\title{Merge Samples}
\usage{
merge_chains(emc)
}
\arguments{
\item{emc}{An emc object, commonly the output of \code{fit()}}
}
\value{
An unlisted emc object with all chains merged
}
\description{
Merges samples from all chains as one unlisted object.
}
\details{
Note that all sampling stages are included in the merged output,
including iterations from the \code{preburn}, \code{burn}, and \code{adapt} stages.
\code{merge_chains(emc)$samples$stage} shows the corresponding sampling stages.
}
