% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{merge_samples}
\alias{merge_samples}
\title{Merge samples}
\usage{
merge_samples(samplers)
}
\arguments{
\item{samplers}{A list of samplers}
}
\value{
an unlisted sampler with all chains merged
}
\description{
Merges samples from all chains together as one unlisted object
}
