% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LBA.R
\name{Mlba}
\alias{Mlba}
\title{Mlba_B
LBA model accommodating missing values (truncation and censoring) and
assuming positive rates (i.e., no intrinsic omissions)}
\usage{
Mlba()
}
\description{
Mlba_B
LBA model accommodating missing values (truncation and censoring) and
assuming positive rates (i.e., no intrinsic omissions)
}
