% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_lnr.R
\name{MlnrMS}
\alias{MlnrMS}
\title{lnrMS Model accommodating missing values (truncation and censoring)}
\usage{
MlnrMS()
}
\value{
A model list with all the necessary functions to sample
}
\description{
lnrMS Model accommodating missing values (truncation and censoring)
}
