% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LBA.R
\name{MUlbaB}
\alias{MUlbaB}
\title{MUlbaB}
\usage{
MUlbaB()
}
\value{
A model list with all the necessary functions to sample
}
\description{
LBA model accommodating missing values (truncation and censoring) and
assuming unbounded rates (i.e., allows intrinsic omissions)
}
