% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_multinomial.R
\name{multinomial_logit}
\alias{multinomial_logit}
\title{Multinomial Logit Response Model}
\usage{
multinomial_logit()
}
\value{
A model list with all the necessary functions to sample
}
\description{
Unordered choice among response alternatives using a softmax over latent
utilities.
}
\details{
Model parameters are:
utility (unbounded).
}
\examples{
dmnl <- design(
  Rlevels = c("left", "right"),
  factors = list(subjects = 1, S = c("left", "right")),
  formula = list(utility ~ 0 + S),
  matchfun = function(d) d$S == d$lR,
  model = multinomial_logit
)
}
