% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_multinomial.R
\name{multinomial_probit}
\alias{multinomial_probit}
\title{Multinomial Probit Response Model}
\usage{
multinomial_probit()
}
\value{
A model list with all the necessary functions to sample
}
\description{
Unordered choice among response alternatives using latent Gaussian utility
noise. Assumes independent Gaussian shocks across alternatives,
which implies correlated pairwise differences in the implied comparison
space.
}
\details{
Model parameters are:
utility (unbounded).
}
\examples{
dmnp <- design(
  Rlevels = c("left", "right"),
  factors = list(subjects = 1, S = c("left", "right")),
  formula = list(utility ~ 0 + S),
  matchfun = function(d) d$S == d$lR,
  model = multinomial_probit
)
}
