% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{p_test}
\alias{p_test}
\title{Title}
\usage{
p_test(
  x,
  x_name = NULL,
  x_fun = NULL,
  fun_name = "fun",
  y = NULL,
  y_name = NULL,
  y_fun = NULL,
  mapped = FALSE,
  x_subject = NULL,
  y_subject = NULL,
  mu = 0,
  alternative = c("less", "greater")[1],
  probs = c(0.025, 0.5, 0.975),
  digits = 2,
  p_digits = 3,
  print_table = TRUE,
  filter = "sample",
  selection = "mu",
  subfilter = 0
)
}
\arguments{
\item{subfilter}{}
}
\description{
Title
}
