% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_CDM.R
\name{PHSDM}
\alias{PHSDM}
\title{The Projected Hyper-Spherical Diffusion Model (PHSDM)}
\usage{
PHSDM()
}
\value{
A model list with all the necessary functions for EMC2 to sample.
}
\description{
Model file to estimate the Projected Hyper-Spherical Diffusion Model (PHSDM) in EMC2.
}
\details{
Parameters and defaults are on the transformed scale.
All parameters not specified in the \code{formula} of \code{design()} are
assumed constant with the defaults in \code{p_types}.\tabular{lllll}{
   Parameter \tab Transform \tab Natural scale \tab Default \tab Interpretation \cr
   v \tab log \tab [0, Inf) \tab log(1) \tab Drift magnitude \cr
   theta1 \tab pnorm \tab [0, 1) \tab qnorm(.5) \tab Drift angle, mapped to 0..pi \cr
   theta2 \tab pnorm \tab [0, 1) \tab qnorm(.5) \tab Drift angle, mapped to 0..pi \cr
   a \tab log \tab [0, Inf) \tab log(1) \tab Boundary radius \cr
   t0 \tab log \tab [0, Inf) \tab log(0) \tab Non-decision time \cr
   s \tab log \tab [0, Inf) \tab log(1) \tab Diffusion scale \cr
   sv \tab log \tab [0, Inf) \tab log(0) \tab Drift SD across trials \cr
}


The response variables are \code{rt}, \code{R}, and \code{R2}.
This implementation assumes fixed boundaries (no decay/collapsing boundary parameter).
}
