% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_funcs.R
\name{plot.emc}
\alias{plot.emc}
\title{Plot function for EMC2 samplers objects}
\usage{
\method{plot}{emc}(x, filter = "sample", selection = c("mu", "sigma2", "alpha"), ...)
}
\arguments{
\item{x}{An object of class \code{emc}}

\item{filter}{A character string indicating the sampling stage to be summarized.
Can be \code{preburn}, \code{burn}, \code{adapt}, or \code{sample}.}

\item{selection}{A character string indicating the parameter group.
Defaults to \code{mu}, \code{sigma2}, and \code{alpha}.}

\item{...}{Optional arguments that can be passed to \code{as_mcmc_new} or \code{plot.default} (see \code{par()})}
}
\description{
Plots posterior densities and traceplots for EMC2 model parameters.
}
\details{
Note that if \code{subfilter} is set to an integer other than 0 (i.e., the default), all iterations
up to that integer are ignored. If a vector is supplied, only those iterations
are considered.
}
