% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_alpha_recovery}
\alias{plot_alpha_recovery}
\title{Title}
\usage{
plot_alpha_recovery(
  tabs,
  layout = c(2, 3),
  do_ci = TRUE,
  ci_col = "grey",
  cap = 0.05,
  do_rmse = FALSE,
  rmse_pos = "topleft",
  rmse_digits = 3,
  pearson_digits = 2,
  do_coverage = FALSE,
  coverage_pos = "bottomright",
  coverage_digits = 1,
  spearman_digits = 2
)
}
\arguments{
\item{spearman_digits}{}
}
\description{
Title
}
