% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_chains}
\alias{plot_chains}
\title{Plot MCMC chains}
\usage{
plot_chains(
  samplers,
  layout = NA,
  subject = NA,
  ylim = NULL,
  selection = "mu",
  filter = "sample",
  thin = 1,
  subfilter = 0,
  plot_acf = FALSE
)
}
\arguments{
\item{samplers}{An EMC2 samplers object}

\item{layout}{A vector specifying the layout as in \code{par(mfrow = layout)}.
If \code{NA}, (i.e., the default) the \code{coda} defaults are used (unless \code{plot_acf = TRUE}),
if \code{NULL} the current plot layout is used.}

\item{subject}{Integer or character vector. Only applicable if \code{selection = "alpha"}. Will plot only these subject(s).
\code{NA} (i.e., the default) will plot all individuals.}

\item{ylim}{A vector. The y-limits of the chain plot.}

\item{selection}{A Character string. Indicates which parameter type to plot (\code{alpha}, \code{mu}, \code{variance}, \code{covariance}, \code{correlation}).
\code{LL} will plot the log-likelihood chains.}

\item{filter}{A string. Specifies from which stage you want to plot the MCMC chains (\code{preburn}, \code{burn}, \code{adapt}, \code{sample})}

\item{thin}{An integer. Will keep only iterations of the MCMC chains that are a multiple of \code{thin}.}

\item{subfilter}{Integer or numeric vector. If an integer is supplied, iterations
up until that integer within the sampling stage \code{filter} are kept. If a vector is supplied, the iterations
within the range are kept.}

\item{plot_acf}{Boolean. If \code{TRUE}, the autocorrelation of the first chain is plotted
(of the -- per default -- three chains).}
}
\value{
None
}
\description{
Plots the trace plots of the MCMC chains on top of each other. Visualizes convergence
and chain stability.
}
\examples{
\dontrun{
# For a set of samplers run using run_emc:
plot_chains(samplers)
# Or for the second subject:
plot_chains(samplers, subject = 2)

# Can also plot the autocorrelation of for example the group-level mean:
plot_chains(samplers, selection = "mu", plot_acf = TRUE)
}
}
