% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_density}
\alias{plot_density}
\title{Plot Defective Densities}
\usage{
plot_density(
  input,
  post_predict = NULL,
  prior_predict = NULL,
  subject = NULL,
  quants = c(0.025, 0.975),
  functions = NULL,
  factors = NULL,
  defective_factor = "R",
  n_cores = 1,
  n_post = 50,
  layout = NA,
  to_plot = c("data", "posterior", "prior")[1:2],
  use_lim = c("data", "posterior", "prior")[1:2],
  remove_na = TRUE,
  legendpos = c("topright", "top"),
  posterior_args = list(),
  prior_args = list(),
  ...
)
}
\description{
Plots panels that contain a set of densities for each level of the specified defective factor in the data.
These densities are defective; their areas are relative to the respective
proportions of the defective factor levels. Across all levels, the area sums to
the probability of a finite RT (per panel, and per context group when the
defective factor is a combination of a stimulus factor and a response factor).
Optionally, posterior/prior predictive densities can be overlaid.
}
\examples{
# Plot defective densities for each subject and the factor combination in the design:
plot_density(forstmann)
# or for one subject:
plot_density(forstmann, subject = 1)
# Now collapsing across subjects and using a different defective factor:
plot_density(forstmann, factors = "S", defective_factor = "E")
# Or plot posterior predictives
plot_density(samples_LNR, n_post = 10)
}
