% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R, R/priors.R, R/s3_funcs.R
\name{plot_design.emc.design}
\alias{plot_design.emc.design}
\alias{plot_design.emc.prior}
\alias{plot_design}
\alias{plot_design.emc}
\title{Plot Design}
\usage{
\method{plot_design}{emc.design}(
  x,
  data = NULL,
  factors = NULL,
  plot_factor = NULL,
  n_data_sim = 10,
  p_vector = NULL,
  functions = NULL,
  ...
)

\method{plot_design}{emc.prior}(
  x,
  data = NULL,
  factors = NULL,
  plot_factor = NULL,
  n_data_sim = 10,
  p_vector = NULL,
  functions = NULL,
  ...
)

plot_design(
  x,
  data = NULL,
  factors = NULL,
  plot_factor = NULL,
  n_data_sim = 10,
  p_vector = NULL,
  functions = NULL,
  ...
)

\method{plot_design}{emc}(
  x,
  data = NULL,
  factors = NULL,
  plot_factor = NULL,
  n_data_sim = 10,
  p_vector = NULL,
  functions = NULL,
  ...
)
}
\arguments{
\item{x}{An \code{emc} or \code{emc.prior} object containing the design to plot}

\item{data}{Optional data to overlay on the design plot}

\item{factors}{Factors to use for varying parameters}

\item{plot_factor}{Optional. Make separate plots for each level of this factor}

\item{n_data_sim}{If data is provided, number of simulated datasets to generate for the plot. Default is 10.}

\item{p_vector}{Only needed when x is an \code{emc.design} object, which parameters to use for data generation.}

\item{functions}{A named list of functions that create additional columns in the data.}

\item{...}{Additional arguments to pass to \code{make_design_plot}}
}
\value{
No return value. Just plots the design
}
\description{
Makes design illustration by plotting simulated data based on the design
}
