% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_fit}
\alias{plot_fit}
\title{Plots overlaying observed and fitted data.}
\usage{
plot_fit(
  data,
  pp,
  subject = NULL,
  factors = NULL,
  functions = NULL,
  stat = NULL,
  stat_name = "",
  adjust = 1,
  quants = c(0.025, 0.5, 0.975),
  do_plot = TRUE,
  xlim = NULL,
  ylim = NULL,
  main = "",
  layout = NULL,
  probs = c(1:99)/100,
  data_lwd = 2,
  fit_lwd = 1,
  qp_cex = 1,
  q_points = c(0.1, 0.3, 0.5, 0.7, 0.9),
  pqp_cex = 0.5,
  lpos = "topleft",
  signalFactor = "S",
  zROC = FALSE,
  qfun = qnorm,
  lim = NULL,
  rocfit_cex = 0.5
)
}
\arguments{
\item{data}{A dataframe. The experimental data in EMC format with at least the subjects factor,
R (response factor) and rt (response time) columns,
and optionally other factor columns of the design.}

\item{pp}{Posterior predictives created by \code{post_predict}}

\item{subject}{Integer or string selecting a subject from the data. If specified will only plot that subject
(default NULL = all).}

\item{factors}{Character vector of factors in data to display separately. If
NULL (default) use names of all columns in data except "trials","R", and "rt".
Omitted factors are aggregated over. If NA treats entire data set as a single cell.
Must be NA or NULL when using stat argument.}

\item{functions}{A named list of functions that create new factors which can then be
used by the factors and stat arguments.}

\item{stat}{A function that takes a the data and returns a single value.}

\item{stat_name}{A string naming what the stat argument calculates.}

\item{adjust}{Numeric. Density function bandwidth adjust parameter. See ``?density`}

\item{quants}{A vector. Quantiles to return when
stat argument is supplied.}

\item{do_plot}{Boolean. Set to \code{FALSE} to only return the quantiles and omit the plots.}

\item{xlim}{x-axis plot limit. If a vector is supplied will use the same axes for all.}

\item{ylim}{y-axis plot limit. If a vector is supplied will use the same axes for all.}

\item{main}{Text title, pasted before cell name.}

\item{layout}{A vector specifying the layout as in par(mfrow = layout).
If NA or NULL use current.}

\item{probs}{Vector of probabilities at which to calculate cdf}

\item{data_lwd}{Integer. Line width for data in cdf}

\item{fit_lwd}{Integer. Line width for posterior predictives in cdf}

\item{qp_cex}{Numeric. Cex for data quantile points in cdf}

\item{q_points}{Vector. Quantile points to plot in cdf}

\item{pqp_cex}{Numeric. Cex for predicted quantile points in cdf}

\item{lpos}{Character. Legend position, see \code{?legend}.}

\item{signalFactor}{The name of the "signal" factor in an ROC.}

\item{zROC}{Boolean. Wheter to plot ROC on transformed scale.}

\item{qfun}{A function. Scale transform function for zROC (default qnorm)}

\item{lim}{Vector. Limits for x and y limit of ROC}

\item{rocfit_cex}{Numeric. Cex for predicted ROC points}
}
\value{
If stat argument is provided a table of observed values and predicted quantiles
}
\description{
If rt  available plots defective cumulative distributions
functions (CDFs), if not plots region under curve (ROC).
CDFs plot the probability of a response, p(R) as a function of response time
(RT) for data (black lines and points at qpoints quantiles) and posterior-
predictive simulations (grey lines and points). Large grey points show the
average of replicate quantiles and small grey points percentiles for
individual replicates, providing a representation of uncertainty in the model
predictions.
}
\details{
If the stat argument (which calculates
a statistic based on the data) is supplied fit is plotted as a density with
a vertical line at the position of the data statistic. If more than one
subject is included data and fits are aggregated over subjects. If data or
fit contains NA in responses or rt, or is.infinite(rt) these are treated as
missing and defective cdfs sum to the probability of non-missing.
}
