% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_fits}
\alias{plot_fits}
\title{Convenience version of plot_fit that automatically applies it to each subject
separately.}
\usage{
plot_fits(
  data,
  pp,
  factors = NULL,
  stat = NULL,
  stat_name = "",
  do_plot = TRUE,
  xlim = NULL,
  ylim = NULL,
  layout = NULL,
  mfcol = TRUE,
  probs = c(1:99)/100,
  data_lwd = 2,
  fit_lwd = 1,
  qp_cex = 1,
  q_points = c(0.1, 0.3, 0.5, 0.7, 0.9),
  pqp_cex = 0.5,
  lpos = "topleft",
  signalFactor = "S",
  zROC = FALSE,
  qfun = NULL,
  lim = NULL,
  rocfit_cex = 0.5
)
}
\arguments{
\item{data}{Data frame with subjects and R factors, and possibly other factors
and an rt column}

\item{pp}{Posterior predictives created by post_predict}

\item{factors}{Character vector of factors in data to display separately. If
NULL (default) use names of all columns in data except "trials","R", and "rt".
Omitted factors are aggregated over.}

\item{stat}{A function that takes a the data and returns a single value.}

\item{stat_name}{A string naming what the stat argument calculates.}

\item{do_plot}{Boolean (default TRUE) for making a plot}

\item{xlim}{x-axis plot limit, 2-vector (same for all) or matrix (one row for each paramter)}

\item{ylim}{y-axis plot limit, 2-vector (same for all) or matrix (one row for each paramter)}

\item{layout}{2-vector specifying par(mfrow) or par(mfcol) (default NULL use current).}

\item{mfcol}{Boolean, default TRUE use mfcol else mfrow.}

\item{probs}{Vector of probabilities at which to calculate cdf (default percentiles)}

\item{data_lwd}{Integer line width for data in cdf (default = 2)}

\item{fit_lwd}{Integer line width for fit in cdf (default = 1)}

\item{qp_cex}{cex for data quantile points in cdf}

\item{q_points}{Quantile points to plot in cdf (default c(.1,.3,.5,.7,.9))}

\item{pqp_cex}{cex for predicted quantile points in cdf}

\item{lpos}{Legend position (see legend)}

\item{signalFactor}{The name of the "signal" factor in an ROC.}

\item{zROC}{Boolean, (default FALSE) to plot ROC on transformed scale.}

\item{qfun}{Scale transform function for zROC (default qnorm)}

\item{lim}{2-vector for x and y limit of ROC}

\item{rocfit_cex}{cex for predicted ROC points.}
}
\description{
Convenience version of plot_fit that automatically applies it to each subject
separately.
}
