% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_mcmc_list}
\alias{plot_mcmc_list}
\title{Plot MCMC.list}
\usage{
plot_mcmc_list(emc, selection = "mu", fun = "traceplot", layout = NA, ...)
}
\arguments{
\item{emc}{An emc object}

\item{selection}{A Character string. Indicates which parameter type to plot (e.g., \code{alpha}, \code{mu}, \code{sigma2}, \code{correlation}).}

\item{fun}{A coda plot function choice from}

\item{layout}{A vector indicating which layout to use as in par(mfrow = layout). If NA, will automatically generate an appropriate layout.}

\item{...}{Optional arguments that can be passed to \code{get_pars},
the chosen coda plot function, or \code{plot.default} (see \code{par()})}
}
\value{
A coda plot
}
\description{
Uses the coda plot functions that are applied across chain
}
