% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{plot_prior}
\alias{plot_prior}
\title{plot_prior
Plots prior distributions simulated by get_prior functions.}
\usage{
plot_prior(
  prior,
  type = NULL,
  add_density = FALSE,
  adjust = 1,
  breaks = 50,
  layout = c(3, 3),
  upper = NULL,
  xlim = NULL,
  mapped = TRUE,
  design = NULL
)
}
\arguments{
\item{prior}{A list of prior samples}

\item{type}{Select type of prior to plot (e.g., "alpha", "mu" etc.), default NULL plots all types present}

\item{add_density}{Boolean (default FALSE) draw density over prior histograms}

\item{adjust}{density smoothing adjustment (Default 1, see density)}

\item{breaks}{Integer number of breaks in histograms}

\item{layout}{par(mfrow) setting (default c(3,3))}

\item{upper}{Vector with parameter names of upper quantile limit of values plotted (default 0.999)}

\item{xlim}{List with parameter names of plot x limits}

\item{mapped}{Boolean (default TRUE) mapped mu and alpha to design on natural scale}

\item{design}{Design corresponding to prior, must be supplied if mapped=TRUE}
}
\description{
plot_prior
Plots prior distributions simulated by get_prior functions.
}
