% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{plot_relations}
\alias{plot_relations}
\title{Plot relations}
\usage{
plot_relations(
  samplers = NULL,
  filter = "sample",
  plot_cred = TRUE,
  plot_means = TRUE,
  only_cred = FALSE,
  nice_names = NULL,
  ...
)
}
\arguments{
\item{samplers}{An EMC2 samplers object, commonly the output of \code{run_emc()}.}

\item{filter}{Character. The stage from which to take the samples, defaults to
the sampling stage \code{sample}.}

\item{plot_cred}{Boolean. Whether to plot the credible intervals or not}

\item{plot_means}{Boolean. Whether to plot the means or not}

\item{only_cred}{Boolean. Whether to only plot credible values}

\item{nice_names}{Character string. Alternative names to give the parameters}

\item{...}{Optional additional arguments}
}
\description{
An adjusted version of the \code{corrplot} package function \code{corrplot()} tailored
to \code{EMC2} and the plotting of correlations.
}
\examples{
\dontrun{
# For a given set of hierarchical factor model samples we can make a
# loadings matrix plot.
plot_relations(samplers, only_cred = TRUE, plot_cred = TRUE)
# We can also plot the correlations, this works for any model that included correlations
plot_relations(samplers, do_corr = T, plot_means = T, only_cred = T)
}
}
