% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{post_predict}
\alias{post_predict}
\title{Generate posterior predictives}
\usage{
post_predict(
  samplers,
  hyper = FALSE,
  n_post = 100,
  filter = "sample",
  subfilter = 0,
  thin = 1,
  n_cores = 1,
  stat = c("random", "mean", "median")[1],
  ...
)
}
\arguments{
\item{samplers}{An EMC2 samplers object from which posterior predictives should
be generated}

\item{hyper}{Boolean. Defaults to \code{FALSE}. If \code{TRUE}, simulates from the group-level (\code{hyper})
parameters instead of the subject-level parameters.}

\item{n_post}{Integer. Number of generated datasets}

\item{filter}{Character. From which sampling stage should the samples be taken?
Defaults to \code{sample}. Choice of the stages \code{preburn}, \code{burn}, \code{adapt}, \code{sample}}

\item{subfilter}{Integer or numeric vector. If integer, will filter out the
supplied number of samples, within \code{filter}. If a numeric vector is supplied,
the supplied range is selected within \code{filter}}

\item{thin}{Integer. By how much should the chains be thinned before simulating from them? Will keep 1/thin samples.}

\item{n_cores}{Integer. Number of cores across which there should be parallellized}

\item{stat}{Character. Can be \code{mean}, \code{median} or \code{random} (i.e., the default).
Will take either random samples from the chain(s) or use the mean or median of the parameter estimates.}

\item{...}{Optional additional arguments}
}
\value{
A list of simulated data sets of length \code{n_post}
}
\description{
Simulate \code{n_post} data sets from the posterior predictive distribution.
}
\examples{
\dontrun{
# based on a set of samplers ran by run_emc we can generate posterior predictives
post_predict(samplers, n_cores = 8)
}
}
