% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{posterior_summary_new}
\alias{posterior_summary_new}
\title{Posterior quantiles}
\usage{
posterior_summary_new(
  emc,
  selection = "mu",
  probs = c(0.025, 0.5, 0.975),
  digits = 3,
  ...
)
}
\arguments{
\item{emc}{An emc object}

\item{selection}{A Character vector. Indicates which parameter types to check (e.g., \code{alpha}, \code{mu}, \code{sigma2}, \code{correlation}).}

\item{probs}{A vector. Indicates which quantiles to return from the posterior.}

\item{digits}{Integer. How many digits to round the quantiles to in the output.}

\item{...}{Optional additional arguments that can be passed to \code{as_mcmc_new}}
}
\value{
A list of posterior quantiles for each parameter group in the selected parameter type.
}
\description{
Returns the quantiles of the selected parameter type.
Full range of possible samples manipulations described in \code{as_mcmc_new}.
}
\examples{
posterior_summary_new(samplers_LNR)
}
