% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_help}
\alias{prior_help}
\title{Prior Specification Information}
\usage{
prior_help(type)
}
\arguments{
\item{type}{A character string indicating which 'type' of model to run (e.g. 'standard' or 'single')}
}
\value{
Invisible return with a list of all the information that is also printed
}
\description{
Prints information associated with the prior for certain 'type'
}
\examples{
prior_help('diagonal')
}
