% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_SDT.R
\name{probit}
\alias{probit}
\title{Gaussian Signal Detection Theory Model}
\usage{
probit()
}
\value{
A model list with all the necessary functions to sample
}
\description{
Discrete choice based on continuous Gaussian latent, with no rt. Model
parameters are mean (unbounded) sd (log scale) and threshold, with a first
value is  on the natural scale, and others for designs with with more than
two responses are threshold increases on a log scale to enforce monotonic
increase on the natural scale.
}
