% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_RDM.R
\name{RDMSWTN_TTcorr}
\alias{RDMSWTN_TTcorr}
\title{Correlated Time-Changed RDMSWTN Race}
\usage{
RDMSWTN_TTcorr(posdrift = TRUE)
}
\arguments{
\item{posdrift}{Logical. Active nonzero \code{rho} requires \code{TRUE}.}
}
\value{
A model list compatible with \code{\link[=design]{design()}}.
}
\description{
Convenience constructor for \code{RDMSWTN_TT(correlated = TRUE)}.
}
