% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{recovery_plot}
\alias{recovery_plot}
\title{Plots recovery of parameters}
\usage{
recovery_plot(
  tabs,
  layout = c(2, 3),
  do_ci = TRUE,
  ci_col = "grey",
  cap = 0.05,
  correlation = "pearson",
  cor_pos = "topleft",
  stat = "coverage",
  stat_pos = "bottomright",
  digits = 3
)
}
\arguments{
\item{tabs}{Tables of actual and estimated parameters (with CIs) from plot_pars}

\item{layout}{A vector specifying the layout as in \code{par(mfrow = layout)}}

\item{do_ci}{Boolean. Add credible intervals to plot?}

\item{ci_col}{String. Color of credible interval}

\item{cap}{Numeric. Width of credible interval cap (passed to \code{arrows})}

\item{correlation}{String. \code{pearson} or \code{spearman}. Which correlation included in the plot}

\item{cor_pos}{String. Position of the correlation (passed to \code{legend})}

\item{stat}{String. \code{coverage} or \code{rmse}. Which statistic included in the plot}

\item{stat_pos}{String. Position of the stat (passed to \code{legend})}

\item{digits}{Integer. Number of digits included in the plot for the correlation and stat.}
}
\value{
Invisible list with RMSE, coverage and Pearson and Spearman correlations.
}
\description{
Uses output from \code{plot_pars} to plot generating vs. estimated (median with CI)
parameters for a simulation study.
}
\examples{
# Make up some values that resemble our posteriors
# Normally this would be true values that were used to simulate the data
pmat <- matrix(rnorm(12, mean = c(-1, -.6, -.4, -1.5), sd = .01), ncol = 4, byrow = TRUE)
# This matrix needs to have the appropriate names
# Conventionally this would be created before make data with true values
colnames(pmat) <- c("m", "m_lMd", "s", "t0")
rownames(pmat) <- c("as1t", "bd6t", "bl1t")
# First use plot_pars
recovered_alpha <- plot_pars(samplers_LNR, selection = "alpha", pars = pmat, do_plot = FALSE)
# Now use the output to make the recovery plots
recovery_plot(recovered_alpha, correlation = "pearson", stat = "rmse")

# Similarly we can plot other group-level parameters
recovered_mu <- plot_pars(samplers_LNR, selection = "mu", pars = pmat[1,], do_plot = FALSE)
recovery_plot(recovered_mu, correlation = "spearman", stat = "coverage")
}
