% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{reduce_samples}
\alias{reduce_samples}
\title{reduce_samples}
\usage{
reduce_samples(samplers, thin = 1, remove = NULL)
}
\arguments{
\item{samplers}{A list of pmwg objects}

\item{thin}{thinning interval}

\item{remove}{integer specifiying initial number of samples to remove}
}
\value{
samplers object trimmed and thinned
}
\description{
Enables removal of initial iterations in sample stage and thinning.
}
