% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{remove_iterations}
\alias{remove_iterations}
\title{remove iterations}
\usage{
remove_iterations(
  pmwg,
  select,
  remove = TRUE,
  last_select = FALSE,
  filter = NULL
)
}
\arguments{
\item{pmwg}{A samplers object}

\item{select}{If scalar remove 1:select, if a vector remove those iterations}

\item{remove}{logical, default TRUE, if remove=FALSE instead keeps these iterations}

\item{last_select}{default FALSE, if TRUE apply select from last iteration}

\item{filter}{stage from which to remove samples, default "sample", if NULL
remove absolute iteration number}
}
\value{
samplers object
}
\description{
remove iterations
}
