% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnr.R
\name{rLNR}
\alias{rLNR}
\title{rLNR()}
\usage{
rLNR(lR, pars, p_types = c("m", "s", "t0"))
}
\arguments{
\item{lR}{empty latent response factor lR with one level for each accumulator.}

\item{pars}{matrix of corresponding parameter values named as in p_types}

\item{p_types}{}
}
\description{
rLNR()
}
