% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{rotate_loadings}
\alias{rotate_loadings}
\title{Rotate loadings based on posterior median}
\usage{
rotate_loadings(emc, rot_fun)
}
\arguments{
\item{emc}{An 'emc' object containing factor analysis results}

\item{rot_fun}{a rotation function for factor loadings, see also \code{GPArotation}W}
}
\value{
An 'emc' object with rotated factor loadings
}
\description{
This function rotates factor loadings using a rotation function
based on the posterior median.
}
