% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{run_samplers}
\alias{run_samplers}
\title{Generic function to run samplers for any stage and any requirements.
Used by run_emc, auto_burn, run_adapt and run_sample
Will break if you skip a stage, the stages have to be run in order (preburn, burn, adapt, sample)
Either iter, max_gd, min_es or min_unique has to be specified. Multiple conditions for finishing can be specified. Will finish if all conditions are met.}
\usage{
run_samplers(
  samplers,
  stage,
  iter = NULL,
  max_gd = NULL,
  min_es = 0,
  min_unique = 750,
  p_accept = 0.8,
  step_size = 100,
  verbose = F,
  autosave = F,
  fileName = NULL,
  particles = NULL,
  particle_factor = 50,
  cores_per_chain = 1,
  cores_for_chains = length(samplers),
  max_trys = 50
)
}
\arguments{
\item{samplers}{A list of samplers, could be in any stage, as long as they've been initialized with make_samplers}

\item{stage}{A string. Indicates which stage is to be run, either preburn, burn, adapt or sample}

\item{iter}{An integer. Indicates how many iterations to run,}

\item{max_gd}{A double. The maximum gelman diagnostic convergence allowed. Will stop if below this number.}

\item{min_es}{An integer. The minimal effective size required to stop running.}

\item{min_unique}{An integer. The minimal number of samples required. Only works in adaptation.}

\item{p_accept}{A double. The target acceptance probability of the MCMC process. This will fine tune the width of the search space. Default = .8}

\item{step_size}{An integer. After each of these steps, the requirements will be checked if they are met and proposal distributions will be updated. Default = 100.}

\item{verbose}{Logical. Whether to print sampling related messages}

\item{autosave}{Logical. Whether to automatically save the samplers to fileName. fileName has to be specified.}

\item{fileName}{A string. Indicates what to save the samplers as, if autosave is enabled.}

\item{particles}{An integer. How many particles to use, default is NULL and particle_factor is used. If specified will override particle_factor}

\item{particle_factor}{An integer. Particle factor multiplied by the square root of the number of sampled parameters will determine the number of particles used.}

\item{cores_per_chain}{An integer. How many cores to use per chain. Parallelizes across participant calculations.}

\item{cores_for_chains}{An integer. How many cores to use across chains. Default is the number of chains.}

\item{max_trys}{An integer. How many times it will try to meet the finish conditions. Default is 50.}
}
\value{
A list of samplers
}
\description{
Generic function to run samplers for any stage and any requirements.
Used by run_emc, auto_burn, run_adapt and run_sample
Will break if you skip a stage, the stages have to be run in order (preburn, burn, adapt, sample)
Either iter, max_gd, min_es or min_unique has to be specified. Multiple conditions for finishing can be specified. Will finish if all conditions are met.
}
