% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{run_stages}
\alias{run_stages}
\title{Title}
\usage{
run_stages(
  sampler,
  iter = c(300, 0, 0),
  verbose = FALSE,
  verbose_run_stage = FALSE,
  particles = NA,
  particle_factor = 50,
  p_accept = NULL,
  n_cores = 1,
  epsilon = NULL,
  start_mu = NULL,
  start_var = NULL,
  mix = NULL,
  epsilon_upper_bound = 15,
  eff_var = NULL,
  eff_mu = NULL
)
}
\arguments{
\item{eff_mu}{}
}
\description{
Title
}
