% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBC.R
\name{SBC_single}
\alias{SBC_single}
\title{Simulation-based calibration}
\usage{
SBC_single(
  design_in,
  prior_in,
  replicates = 250,
  trials = 100,
  plot_data = FALSE,
  verbose = TRUE,
  n_post = 1000,
  fileName = NULL,
  ...
)
}
\arguments{
\item{design_in}{An emc design list. The design of the model to be used in SBC}

\item{prior_in}{An emc prior list. The prior for the design to be used in SBC}

\item{replicates}{Integer. The number of samples to draw from the prior}

\item{trials}{Integer. The number of trials of the simulated data (per subject)}

\item{plot_data}{Boolean. Whether to plot the data simulated (aggregated across subjects)}

\item{verbose}{Verbose. Whether to print progress related messages}

\item{n_post}{Integer. The number of posterior samples to be taken and calculated the rank across}

\item{fileName}{Character. Highly recommended, saves temporary results to the fileName}

\item{...}{A list of optional additional arguments that can be passed to \code{fit} and \code{make_emc}}
}
\value{
A list with the rank of the subject-level parameters and the prior-samples
}
\description{
Runs single subject SBC for an EMC2 model and associated design. Note that the
rank of the subject-level parameters is returned.
}
