% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{select}
\alias{select}
\title{select}
\usage{
select(
  pmwgs,
  keep,
  keep_last = FALSE,
  filter = "sample",
  only_keep_filter = TRUE
)
}
\arguments{
\item{pmwgs}{An emc object}

\item{keep}{If scalar select 1:keep, if a vector select those iterations}

\item{keep_last}{default FALSE, if TRUE apply keep from last iteration}

\item{filter}{the stage from which to select samples, default "sample",
if NULL select on absolute iteration number (including "init" which is iteration 1)}

\item{only_keep_filter}{logical default TRUE, keep samples only from
filter stage, empty all other stages, if FALSE apply selection only to stage
specified in filter, keep all other staged (ignored if filter=NULL).}
}
\value{
An emc object
}
\description{
Selects iterations of an emc object to keep.
}
