% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{set_custom_kernel_pointers}
\alias{set_custom_kernel_pointers}
\title{(Re-)Set pointers of custom C++ trend kernels to an emc object}
\usage{
set_custom_kernel_pointers(emc, ptrs)
}
\arguments{
\item{emc}{An emc object}

\item{ptrs}{A list of pointers, as generated by \code{get_trend_pointers}}
}
\value{
An emc object with the custom pointers re-instated.
}
\description{
When an emc object is loaded from disk, or returned by forked processes, the pointers to custom kernels
need to be re-created. This is a convenience function to do this.
}
\examples{
\donttest{
pointers <- get_custom_kernel_pointers(emc)
emc <- set_custom_kernel_pointers(emc, pointers)
}
}
