% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_SS.R
\name{SSD_function}
\alias{SSD_function}
\title{Assign Stop-Signal Delays (SSDs) to trials}
\usage{
SSD_function(d, SSD = NA, pSSD = 0.25)
}
\arguments{
\item{d}{A data frame containing trial data with a response factor column 'lR'}

\item{SSD}{A vector of stop-signal delays to be assigned to trials}

\item{pSSD}{A vector of probabilities for each SSD value. If length is one less than SSD,
the remaining probability is calculated automatically. Default is 0.25.}
}
\value{
A vector of SSDs with the same length as the number of rows in 'd'.
Trials without a stop signal are assigned Inf.
}
\description{
This function assigns stop-signal delays (SSDs) to trials based on specified probabilities.
}
\examples{
# data <- data.frame(lR = factor(c("left", "right")), rt = c(0.5, 0.6))
# SSD_function(data, SSD = c(0.2, 0.3), pSSD = c(0.2, 0.3))
}
