% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_SS.R
\name{SShybrid}
\alias{SShybrid}
\title{Stop-signal Hybrid (RDM go, ExGaussian stop) race}
\usage{
SShybrid()
}
\value{
A model list with all the necessary functions to sample
}
\description{
Model file to estimate the Hybrid race model for Stop-Signal data
}
\details{
Model files are almost exclusively used in \code{design()}.

Default values are used for all parameters that are not explicitly listed in the \code{formula}
argument of \code{design()}.They can also be accessed with \code{SShybrid()$p_types}.
}
