% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_SS.R
\name{SStexG}
\alias{SStexG}
\title{The ex-Gaussian race model of the stop signal task}
\usage{
SStexG()
}
\value{
A model list with all the necessary functions to sample
}
\description{
Model file to estimate the ex-Gaussian race model of stop signal task data in EMC2.
}
\details{
Model files are almost exclusively used in \code{design()}.

Default values are used for all parameters that are not explicitly listed in the \code{formula}
argument of \code{design()}.They can also be accessed with \code{SSexG()$p_types}.\tabular{llllll}{
   \strong{Parameter} \tab \strong{Transform} \tab \strong{Natural scale} \tab \strong{Default} \tab \strong{Mapping} \tab \strong{Interpretation} \cr
   \emph{mu} \tab log \tab [0, Inf] \tab log(.4) \tab  \tab Mean of Gaussian component of ex-Gaussian go finish time distribution \cr
   \emph{sigma} \tab log \tab [0, Inf] \tab log(.05) \tab  \tab Standard deviation of Gaussian component of ex-Gaussian go finish time distribution \cr
   \emph{tau} \tab log \tab [0, Inf] \tab log(.1) \tab  \tab Mean (inverse rate) of exponential component of ex-Gaussian go finish time distribution \cr
   \emph{muS} \tab log \tab [0, Inf] \tab log(.3) \tab  \tab Mean of Gaussian component of ex-Gaussian stop finish time distribution \cr
   \emph{sigmaS} \tab log \tab [0, Inf] \tab log(.025) \tab  \tab Standard deviation of Gaussian component of ex-Gaussian stop finish time distribution \cr
   \emph{tauS} \tab log \tab [0, Inf] \tab log(.05) \tab  \tab Mean (inverse rate) of exponential component of ex-Gaussian stop finish time distribution \cr
   \emph{tf} \tab probit \tab [0, 1] \tab qnorm(0) \tab  \tab Attentional lapse rate for stop process ("trigger failure") \cr
   \emph{gf} \tab probit \tab [0, 1] \tab qnorm(0) \tab  \tab Attentional lapse rate for go process ("go failure") \cr
   \emph{exg_lb} \tab - \tab [-Inf, Inf] \tab .05 \tab  \tab Lower bound of ex-Gaussian go finish time distribution \cr
   \emph{exgS_lb} \tab - \tab [-Inf, Inf] \tab .05 \tab  \tab Lower bound of ex-Gaussian stop finish time distribution \cr
}


Because the ex-Gaussian stop signal model is a race model, it has one accumulator per response option.
EMC2 automatically constructs a factor representing the accumulators \code{lR} (i.e., the
latent response) with level names taken from the \code{R} column in the data.

For race models, the \code{design()} argument \code{matchfun} can be provided, a
function that takes the \code{lR} factor (defined in the augmented data (d)
in the following function) and returns a logical defining the correct response.
In the example below, the match is simply such that the \code{S} factor equals the
latent response factor: \code{matchfun=function(d)d$S==d$lR}. Then \code{matchfun} is
used to automatically create a latent match (\code{lM}) factor with
levels \code{FALSE} (i.e., the stimulus does not match the accumulator) and \code{TRUE}
(i.e., the stimulus does match the accumulator). This is added internally
and can also be used in model formula.

The race model of the stop signal task was first introduced by Logan & Cowan (1984), who modeled task performance as a race between two stochastic processes: a go process for response execution (triggered by the choice stimulus), and a stop process for response inhibition (triggered by the stop signal).
Several review papers provide further background on the task paradigm and modelling framework (e.g., Matzke et al., 2018; Verbruggen et al., 2019; Colonius et al., 2023).

Early work on the ex-Gaussian distribution as a descriptive model of RT data includes Ratcliff & Murdock (1976), Ratcliff (1979), and Heathcote et al. (1991).

Matzke et al. (2013) introduced the ex-Gaussian distribution as a parametric model of the finish time distributions of the go and stop processes.
This was extended in Matzke et al. (2019) to allow for an arbitrary number of go processes.

Attentional lapse rate parameters for the stop process and go process, termed "trigger failure" and "go failure" respectively, were introduced in Matzke et al. (2017) and Matzke et al. (2019).

Lower truncation of the ex-Gaussian finish time distributions was introduced in Tanis et al. (2024).

Note that the ex-Gaussian parameters do not have unique psychological interpretations (Matzke & Wagenmakers, 2009). Parameter estimates are typically summarised in terms of the mean of the ex-Gaussian distribution, which is given by \code{mu + tau}.
}
\references{
Colonius, H., & Diederich, A. (2023). Modeling response inhibition in the stop signal task. \emph{F. G. Ashby, H. Colonius, & E. N. Dzhafarov (Eds.), New Handbook of Mathematical Psychology}, \emph{3}, 311-356. \doi{10.1017/9781108902724.008}

Heathcote, A., Popiel, S. J., & Mewhort, D. J. (1991). Analysis of response time distributions: an example using the Stroop task. \emph{Psychological Bulletin}, \emph{109}(2), 340. \doi{10.1037/0033-2909.109.2.340}

Logan, G. D., & Cowan, W. B. (1984). On the ability to inhibit thought and action: A theory of an act of control. \emph{Psychological Review}, \emph{91}(3), 295. \doi{10.1037/0033-295X.91.3.295}

Matzke, D., & Wagenmakers, E. J. (2009). Psychological interpretation of the ex-Gaussian and shifted Wald parameters: A diffusion model analysis. \emph{Psychonomic Bulletin & Review}, \emph{16}, 798-817. \doi{10.3758/PBR.16.5.798}

Matzke, D., Dolan, C. V., Logan, G. D., Brown, S. D., & Wagenmakers, E. J. (2013). Bayesian parametric estimation of stop-signal reaction time distributions. \emph{Journal of Experimental Psychology: General}, \emph{142}(4), 1047. \doi{10.1037/a0030543}

Matzke, D., Love, J., & Heathcote, A. (2017). A Bayesian approach for estimating the probability of trigger failures in the stop-signal paradigm. \emph{Behavior Research Methods}, \emph{49}, 267-281. \doi{10.3758/s13428-015-0695-8}

Matzke, D., Verbruggen, F., & Logan, G. (2018). The stop-signal paradigm. \emph{Stevens’ handbook of experimental psychology and cognitive neuroscience}, \emph{5}, 383-427. \doi{10.1002/9781119170174.epcn510}

Matzke, D., Curley, S., Gong, C. Q., & Heathcote, A. (2019). Inhibiting responses to difficult choices. \emph{Journal of Experimental Psychology: General}, \emph{148}(1), 124. \doi{10.1037/xge0000525}

Ratcliff, R., & Murdock, B. B. (1976). Retrieval processes in recognition memory. \emph{Psychological Review}, \emph{83}(3), 190. \doi{10.1037/0033-295X.83.3.190}

Ratcliff, R. (1979). Group reaction time distributions and an analysis of distribution statistics. \emph{Psychological Bulletin}, \emph{86}(3), 446. \doi{10.1037/0033-2909.86.3.446}

Tanis, C. C., Heathcote, A., Zrubka, M., & Matzke, D. (2024). A hybrid approach to dynamic cognitive psychometrics: Dynamic cognitive psychometrics. \emph{Behavior Research Methods}, \emph{56}(6), 5647-5666. \doi{10.3758/s13428-023-02295-y}

Verbruggen, F., Aron, A. R., Band, G. P., Beste, C., Bissett, P. G., Brockett, A. T., ... & Boehler, C. N. (2019). A consensus guide to capturing the ability to inhibit actions and impulsive behaviors in the stop-signal task. \emph{elife}, \emph{8}, e46323. \doi{10.7554/eLife.46323}
}
