% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_SS.R
\name{staircase_fun}
\alias{staircase_fun}
\title{Staircase function}
\usage{
staircase_fun(
  dadm,
  p = 0.25,
  pars = NULL,
  SSD0 = 0.25,
  stairstep = 0.05,
  stairmin = 0,
  stairmax = Inf
)
}
\arguments{
\item{dadm}{A data-augmented design}

\item{p}{Proportion of stop-trials}

\item{pars}{Parameter matrix, needs to have the same number of rows as \code{dadm}}

\item{SSD0}{Minimum SSD in seconds}

\item{stairstep}{Step size in seconds}

\item{stairmin}{Minimum SSD in seconds}

\item{stairmax}{Maximum SSD in seconds}
}
\value{
A data frame with variables \code{R}, \code{rt}, and \code{SSD}
}
\description{
Staircase function for stop-signal models. Can be used to generate
stop-signal data.
}
