% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{standardize_loadings}
\alias{standardize_loadings}
\title{Standardized factor loadings}
\usage{
standardize_loadings(
  samplers = NULL,
  loadings = NULL,
  sig_err_inv = NULL,
  filter = "sample",
  merge_chains = T
)
}
\arguments{
\item{samplers}{A list of samplers from a hierarchical factor analysis model}

\item{loadings}{Array of pars by factors by iters. Can also specify loadings instead of samplers}

\item{sig_err_inv}{Array of pars by iters. Can also specify sig_err_inv instead of samplers}

\item{filter}{Character. From which stage to take samples}

\item{merge_chains}{Return the loadings for each chain separately or merged together.}
}
\value{
standardized loadings
}
\description{
Returns a set of standardized factor loadings.
Standardization with residual error as well as described in Stevenson et al. 2024.
}
