% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{summary.emc.design}
\alias{summary.emc.design}
\title{Summary method for emc.design objects}
\usage{
\method{summary}{emc.design}(object, ...)
}
\arguments{
\item{object}{An object of class \code{emc.design} containing the design to summarize}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the design matrices
}
\description{
Prints a summary of the design object, including sampled parameters and design matrices.
}
