% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_design.R
\name{summary.emc.group_design}
\alias{summary.emc.group_design}
\title{Summary method for emc.group_design objects}
\usage{
\method{summary}{emc.group_design}(object, ...)
}
\arguments{
\item{object}{An object of class \code{emc.group_design} containing the design to summarize}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the design matrices
}
\description{
Prints a summary of the group_design object
For continuous covariates just prints one row, instead of all covariates.
}
