% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{summary.emc.prior}
\alias{summary.emc.prior}
\title{Summary method for emc.prior objects}
\usage{
\method{summary}{emc.prior}(object, ...)
}
\arguments{
\item{object}{An object of class 'emc.prior' containing prior specifications}

\item{...}{Additional arguments passed to other methods (not currently used)}
}
\value{
Invisibly returns NULL. Called for its side effect of printing the summary.
}
\description{
Prints a summary of the prior specification, including descriptions of the prior types
and their associated hyperparameters.
}
\examples{
# Take a prior object
prior <- get_prior(samples_LNR)
summary(prior)

}
\seealso{
\code{\link{prior}} for creating prior objects
}
