% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LBA.R
\name{TC2lbaB}
\alias{TC2lbaB}
\title{Threshold Count 2 LBA model}
\usage{
TC2lbaB()
}
\value{
A model list with all the necessary functions to sample
}
\description{
All thresholds greater than A, DT1 <- A+DT1, b <- DT1 + B
}
