% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MTR.R
\name{TC2PLBA}
\alias{TC2PLBA}
\title{Threshold Count 2 LBA model}
\usage{
TC2PLBA()
}
\value{
A model list with all the necessary functions to sample
}
\description{
Intermediate thresholds and gB proportional to B
}
