% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MTR.R
\name{TC3LBA}
\alias{TC3LBA}
\title{Threshold Count 3 LBA model}
\usage{
TC3LBA()
}
\value{
A model list with all the necessary functions to sample
}
\description{
All thresholds greater than A, DT1 <- A+DT1, DT2 <- DT1 + DT2, b <- DT2 + B
}
