% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LBA_extras.R
\name{TC3lbaB}
\alias{TC3lbaB}
\title{Threshold Count 3 LBA model}
\usage{
TC3lbaB()
}
\value{
A model list with all the necessary functions to sample
}
\description{
All thresholds greater than A, DT1 <- A+DT1, DT2 <- DT1 + DT2, b <- DT2 + B
}
