% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MTR.R
\name{TC3PLBA}
\alias{TC3PLBA}
\title{Threshold count 3 Threshold LBA model}
\usage{
TC3PLBA()
}
\value{
A model list with all the necessary functions to sample
}
\description{
Intermediate thresholds and gB proportional to B, can be less than A
}
