% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MTR.R
\name{TC4LBA}
\alias{TC4LBA}
\title{Threshold Count 4 LBA model}
\usage{
TC4LBA()
}
\value{
A model list with all the necessary functions to sample
}
\description{
All thresholds greater than A, DT1 <- A+DT1, DT2 <- DT1 + DT2, DT3 <- DT2 + DT3, b <- DT3 + B
}
