% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MTR.R
\name{TC4PLBA}
\alias{TC4PLBA}
\title{Threshold count 4 Threshold LBA model}
\usage{
TC4PLBA()
}
\value{
A model list with all the necessary functions to sample
}
\description{
Intermediate thresholds and gb proportional to b, can be less than A
}
