% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{thin_chains}
\alias{thin_chains}
\title{Title}
\usage{
thin_chains(samplers, thin = 5)
}
\arguments{
\item{samplers}{A list of samplers.}

\item{thin}{An integer. Indicates by how much to thin.}
}
\value{
A thinned list of samplers
}
\description{
Title
}
