% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{trend_help}
\alias{trend_help}
\title{Get help information for trend kernels and bases}
\usage{
trend_help(kernel = NULL, base = NULL, show_experimental = FALSE, ...)
}
\arguments{
\item{kernel}{Character string specifying the kernel type to get information about}

\item{base}{Character string specifying the base type to get information about}

\item{show_experimental}{Boolean, if TRUE, will also show information on experimental kernels. These could produce unexpected results or crash. Use at your own risk.}

\item{...}{Additional arguments}
}
\value{
Formatted trend information
}
\description{
Get help information for trend kernels and bases
}
\examples{
# Get information about exponential increasing kernel
trend_help(kernel = "exp_incr")

# Get information about linear base
trend_help(base = "lin")

# Return available kernel and base types
trend_help()
}
