% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{update2version}
\alias{update2version}
\title{Update EMC Objects to the Current Version}
\usage{
update2version(emc)
}
\arguments{
\item{emc}{An EMC object to update}
}
\value{
An updated EMC object compatible with the current version
}
\description{
This function updates EMC objects created with older versions of the package to be compatible with the current version.
}
\examples{
# Update the model to current version
updated_model <- update2version(samples_LNR)

}
