% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{update_model_trend}
\alias{update_model_trend}
\title{Update a model function with trend parameter types and transforms}
\usage{
update_model_trend(trend, model)
}
\arguments{
\item{trend}{An \code{emc2_trend} object created by \code{\link[=make_trend]{make_trend()}}.}

\item{model}{A model function.}
}
\value{
An updated model function.
}
\description{
Update a model function with trend parameter types and transforms
}
