% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{auto_burn}
\alias{auto_burn}
\title{Runs burn-in for samplers.}
\usage{
auto_burn(
  samplers,
  max_gd = 1.2,
  min_es = 0,
  preburn = 150,
  p_accept = 0.8,
  step_size = 100,
  verbose = F,
  autosave = F,
  fileName = NULL,
  particles = NULL,
  particle_factor = 50,
  cores_per_chain = 1,
  cores_for_chains = length(samplers),
  max_trys = 50
)
}
\arguments{
\item{samplers}{A list of samplers, could be in any stage, as long as they've been initialized with make_samplers}

\item{max_gd}{A double. The maximum gelman diagnostic convergence allowed. Will stop if below this number.}

\item{min_es}{An integer. The minimal effective size required.}

\item{preburn}{An integer. The number of iterations run for preburn stage.}

\item{p_accept}{A double. The target acceptance probability of the MCMC process. This will fine tune the width of the search space. Default = .8}

\item{step_size}{An integer. After each of these steps, the requirements will be checked if they are met and proposal distributions will be updated. Default = 100.}

\item{verbose}{Logical. Whether to print sampling related messages}

\item{autosave}{Logical. Whether to automatically save the samplers to fileName. fileName has to be specified.}

\item{fileName}{A string. Indicates what to save the samplers as, if autosave is enabled.}

\item{particles}{An integer. How many particles to use, default is NULL and particle_factor is used. If specified will override particle_factor}

\item{particle_factor}{An integer. Particle factor multiplied by the square root of the number of sampled parameters will determine the number of particles used.}

\item{cores_per_chain}{An integer. How many cores to use per chain. Parallelizes across participant calculations.}

\item{cores_for_chains}{An integer. How many cores to use across chains. Default is the number of chains.}

\item{max_trys}{An integer. How many times it will try to meet the finish conditions. Default is 50.}
}
\value{
A list of samplers
}
\description{
Special instance of \code{run_samplers}, with default arguments specified for completing burn_in.
Will run both preburn and burn.
}
