% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LBA.R
\name{BAwL}
\alias{BAwL}
\title{The Ballistic Accumulator Model with Leakage (modified race-variant of Brown & Heathcote, 2005)}
\usage{
BAwL(posdrift = TRUE)
}
\description{
The Ballistic Accumulator Model with Leakage (modified race-variant of Brown & Heathcote, 2005)
}
