% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{chain_n}
\alias{chain_n}
\title{Returns the number of samples per chain per stage}
\usage{
chain_n(samplers)
}
\arguments{
\item{samplers}{A list of samplers, could be in any stage}
}
\value{
A table of iterations per stage per chain
}
\description{
Returns the number of samples per chain per stage
}
