% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{chain_n}
\alias{chain_n}
\title{Get chain iterations}
\usage{
chain_n(emc)
}
\arguments{
\item{emc}{A list, the output of \code{fit()}.}
}
\value{
A matrix
}
\description{
Returns a matrix with the number of samples per chain for each stage that is present
in the emc object (i.e., \code{preburn}, \code{burn}, \code{adapt},
\code{sample}). The number of rows of the matrix reflects the number of chains
and the number of columns the number of sampling stages.
}
\examples{
chain_n(samples_LNR)

}
