% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{check_run}
\alias{check_run}
\title{Title}
\usage{
check_run(
  samples,
  pdf_name = "check_run.pdf",
  interactive = TRUE,
  filter = "sample",
  subfilter = 0,
  thin = 1,
  layout = c(3, 4),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{height}{}
}
\description{
Title
}
