% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{check_run}
\alias{check_run}
\title{Convergence checks for an EMC2 samplers object}
\usage{
check_run(
  samples,
  pdf_name = "check_run.pdf",
  interactive = TRUE,
  filter = "sample",
  subfilter = 0,
  thin = 1,
  print_IAT = FALSE,
  subject = NULL,
  layout = c(3, 4),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{samples}{A list of samplers or samplers converted to mcmc objects.}

\item{pdf_name}{A character string, indicating the path to and name of the PDF file.}

\item{interactive}{A Boolean, defaults to \code{TRUE}. Indicates whether console output between
hierarchical parameter types should be paused or not.}

\item{filter}{A character string, indicating the sampling stage to examine.
Can be \code{preburn}, \code{burn}, \code{adapt}, or \code{sample}.}

\item{subfilter}{An integer or vector. If an integer is supplied, iterations up until
that integer are examined. If a vector is supplied, everything in that range is examined.}

\item{thin}{An integer. Keeps only iterations that are a multiple of thin.}

\item{print_IAT}{Include a print out of the Integrated Autocorrelation Time.}

\item{subject}{An integer or character string denoting a particular subject.
Creates convergence diagnostics only for that subject within a hierarchical model.}

\item{layout}{A vector specifying the layout as in \code{par(mfrow = layout)}.
If \code{NA} (i.e., the default), it uses \code{coda} defaults, if it is \code{NULL}, it uses
the current layout.}

\item{width}{PDF page width in inches.}

\item{height}{PDF page height in inches.}
}
\value{
A PDF written to \code{pdf_name}.
}
\description{
Runs a series of convergence checks, prints statistics to the console, and
saves traceplots to a PDF file.
}
\details{
Note that the \code{Rhat} is calculated by doubling the number of chains by
first splitting chains into first and second half, so it also a test of
stationarity.

Efficiency of sampling is (optionally) indicated by integrated
autocorrelation time (from the \code{LaplacesDemon} R package) and by the effective
number of samples (from the \code{coda} R package).
}
\examples{
\dontrun{
check_run(samplers)
}
}
