% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{check_trend}
\alias{check_trend}
\title{Check and update formula list for trend parameters}
\usage{
check_trend(
  trend,
  covariates = NULL,
  model = NULL,
  formula = NULL,
  parameter_design = NULL
)
}
\arguments{
\item{trend}{An \code{emc2_trend} object created by \code{\link[=make_trend]{make_trend()}}.}

\item{covariates}{Character vector of covariate column names in the data.}

\item{model}{A model function, or NULL.}

\item{formula}{List of formulas, or NULL.}

\item{parameter_design}{A parameter_design list, or NULL.}
}
\value{
Updated formula list with intercept formulas added for missing trend
parameters.
}
\description{
Check and update formula list for trend parameters
}
