% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{compare}
\alias{compare}
\title{Information criteria and marginal likelihoods}
\usage{
compare(
  sList,
  filter = "sample",
  subfilter = 0,
  use_best_fit = TRUE,
  BayesFactor = TRUE,
  cores_for_props = 4,
  cores_per_prop = 1,
  print_summary = TRUE,
  digits = 0,
  digits_p = 3,
  ...
)
}
\arguments{
\item{sList}{List of samples objects}

\item{filter}{A string. Specifies which stage the samples are to be taken from \code{"preburn"}, \code{"burn"}, \code{"adapt"}, or \code{"sample"}}

\item{subfilter}{An integer or vector. If an integer, it will exclude up until the integer within the filter.}

\item{use_best_fit}{Boolean, defaults to \code{TRUE}, uses the minimal or mean likelihood (whichever is better) in the
calculation, otherwise always uses the mean likelihood.}

\item{BayesFactor}{Boolean, defaults to \code{TRUE}. Include marginal likelihoods as estimated using WARP-III bridge sampling.
Usually takes a minute per model added to calculate}

\item{cores_for_props}{Integer, how many cores to use for the Bayes factor calculation, here 4 is the default for the 4 different proposal densities to evaluate, only 1, 2 and 4 are sensible.}

\item{cores_per_prop}{Integer, how many cores to use for the Bayes factor calculation if you have more than 4 cores available. Cores used will be cores_for_props * cores_per_prop. Best to prioritize cores_for_props being 4 or 2}

\item{print_summary}{Boolean (default \code{TRUE}), print table of results}

\item{digits}{Integer, significant digits in printed table except model weights}

\item{digits_p}{Integer, significant digits in printed table for model weights}

\item{...}{Additional, optional arguments}
}
\value{
Matrix of effective number of parameters, mean deviance, deviance of
mean, DIC, BPIC, Marginal Deviance (if \code{BayesFactor=TRUE}) and associated weights.
}
\description{
Returns the BPIC/DIC or marginal deviance (-2*marginal likelihood) for a list of samples objects.
}
\examples{
\dontrun{
# Define a samplers list of two (or more different models)
# Here the full model is a list of samplers with the hypothesized effect
# The null model is a list of samplers without the hypothesized effect
sList <- list(full_model, null_model)
# By default emc uses 4 cores to parallelize marginal likelihood estimation across proposals
# So cores_per_prop = 3 results in 12 cores used.
compare(sList, cores_per_prop = 3)
}
}
