% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{compare_IC}
\alias{compare_IC}
\title{Title}
\usage{
compare_IC(
  sList,
  filter = "sample",
  subfilter = 0,
  use_best_fit = TRUE,
  print_summary = TRUE,
  digits = 0,
  digits_p = 3,
  subject = NULL
)
}
\arguments{
\item{subject}{}
}
\description{
Title
}
