% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{contr.anova}
\alias{contr.anova}
\title{Anova Style Contrast Matrix}
\usage{
contr.anova(n)
}
\arguments{
\item{n}{An integer. The number of items for which to create the contrast}
}
\value{
A contrast matrix.
}
\description{
Similar to \code{contr.helmert}, but then scaled to estimate differences between conditions. Use in \code{design()}.
}
\examples{
{
design_DDMaE <- design(data = forstmann,model=DDM, contrasts = list(E = contr.anova),
formula =list(v~S,a~E, t0~1, s~1, Z~1, sv~1, SZ~1),
constants=c(s=log(1)))
}
}
