% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{contr.decreasing}
\alias{contr.decreasing}
\title{Contrast to enforce decreasing estimates}
\usage{
contr.decreasing(n)
}
\arguments{
\item{n}{an integer. The number of items for which to create the contrast.}
}
\value{
a contrast matrix.
}
\description{
Each level will be estimated as a reduction from the previous level
}
\examples{
{
design_DDMaE <- design(data = forstmann,model=DDM, contrasts = list(E = contr.decreasing),
formula =list(v~S,a~E, t0~1, s~1, Z~1, sv~1, SZ~1),
constants=c(s=log(1)))
}
}
