% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{contr.increasing}
\alias{contr.increasing}
\title{Contrast Enforcing Increasing Estimates}
\usage{
contr.increasing(n)
}
\arguments{
\item{n}{an integer. The number of items for which to create the contrast.}
}
\value{
a contrast matrix.
}
\description{
Each level will be estimated additively from the previous level
}
\examples{
{
design_DDMaE <- design(data = forstmann,model=DDM, contrasts = list(E = contr.increasing),
formula =list(v~S,a~E, t0~1, s~1, Z~1, sv~1, SZ~1),
constants=c(s=log(1)))
}
}
