% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{contr.increasing}
\alias{contr.increasing}
\title{contr.increasing()}
\usage{
contr.increasing(n, levels = NULL)
}
\arguments{
\item{n}{an integer. The number of items for which to create the contrast.}

\item{levels}{Character vector. the factor levels which will be the colnames of the returning matrix.}
}
\value{
a contrast matrix.
}
\description{
special contrast for SDT threshold factor
first = intercept, cumsum other (positive) levels to force non-decreasing
}
