% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_DDM.R
\name{DDMGNG}
\alias{DDMGNG}
\title{The GNG (go/nogo) Diffusion Decision Model}
\usage{
DDMGNG()
}
\value{
A model list with all the necessary functions to sample
}
\description{
In the GNG paradigm one of the two possible choices results in a response
being withheld (a non-response). A response is represented by a finite
response time and the go response level of \code{R}; a withheld response is
represented by \code{rt = NA} (or, internally during simulation, by an infinite
finishing time) and the nogo level of \code{R}. The paradigm must also specify a
response window, \code{TIMEOUT}, after which withholding is assumed.
}
\details{
The DDM parameterization is the same as \code{\link[=DDM]{DDM()}}, with one addition in the
trial-dependent parameter matrix: \code{TIMEOUT} and the numeric code of \code{Rnogo}
are copied from the augmented data. The likelihood uses the probability of
not reaching the go boundary before \code{TIMEOUT},
\code{1 - P(RT <= TIMEOUT | Rgo)}, for trials recorded as nogo.\tabular{llllll}{
   \strong{Parameter} \tab \strong{Transform} \tab \strong{Natural scale} \tab \strong{Default} \tab \strong{Mapping} \tab \strong{Interpretation} \cr
   \emph{v} \tab identity \tab [-Inf, Inf] \tab 1 \tab  \tab Mean drift rate. \cr
   \emph{a} \tab log \tab [0, Inf] \tab log(1) \tab  \tab Boundary separation. \cr
   \emph{t0} \tab log \tab [0, Inf] \tab log(0) \tab  \tab Non-decision time. \cr
   \emph{s} \tab log \tab [0, Inf] \tab log(1) \tab  \tab Within-trial diffusion scale; conventionally fixed to 1. \cr
   \emph{Z} \tab probit \tab [0, 1] \tab qnorm(.5) \tab  \tab Relative starting point; .5 is unbiased. \cr
   \emph{SZ} \tab probit \tab [0, 1] \tab qnorm(0) \tab  \tab Relative between-trial starting-point range. \cr
   \emph{sv} \tab log \tab [0, Inf] \tab log(0) \tab  \tab Between-trial SD of drift rate. \cr
   \emph{st0} \tab log \tab [0, Inf] \tab log(0) \tab  \tab Between-trial range of non-decision time. \cr
}


The internal DDM parameters are \code{z = Z * a} and
\code{sz = 2 * SZ * min(Z, 1 - Z) * a}. The \code{Rnogo} function must return the
response-factor level that denotes withholding, and \code{Rgo} must return the
corresponding go level. Both functions, together with \code{TIMEOUT}, are
supplied through the \code{functions} argument of \code{\link[=design]{design()}}. Their outputs must
have the same response-factor levels as the data.
As in \code{\link[=DDM]{DDM()}}, optional fitting parameters are \code{pContaminant}, the omission
probability, and \code{pGuess}, the uniform-outlier probability.

The model used is described in the following paper, with the addition of
modeling the TIMEOUT (which is considered but not used in this paper).

Gomez, P., Ratcliff, R., & Perea, M. (2007). A Model of the Go/No-Go Task.
Journal of Experimental Psychology: General, 136(3), 389-413.
https://doi.org/10.1037/0096-3445.136.3.389

To use this model, three functions must be supplied in the design's
\code{functions} argument with the names \code{TIMEOUT}, \code{Rnogo}, and \code{Rgo}. For
example, assuming a 2.5 second timeout, and an \code{R} factor with levels
\code{c("no", "yes")} where \code{"no"} denotes withholding:

TIMEOUT=function(d)rep(2.5,nrow(d))
Rnogo=function(d)factor(rep("no",nrow(d)),levels=c("no","yes"))
Rgo=function(d)factor(rep("yes",nrow(d)),levels=c("no","yes")))

See the help for DDM for further details. At present this model is not fully
implemented in C, so is a little slower to use than the DDM, but not greatly.
}
\examples{
dGNG <- design(Rlevels = c("left","right"),
               factors=list(subjects=1,S=c("left","right")),
               functions=list(
               TIMEOUT=function(d)rep(2.5,nrow(d)),
               # no go response level
               Rnogo=function(d)factor(rep("left",nrow(d)),levels=c("left","right")),
               # go response level
               Rgo=function(d)factor(rep("right",nrow(d)),levels=c("left","right"))),
               formula=list(v~S,a~1, Z~1, t0~1),
               model=DDMGNG)

p_vector <- sampled_pars(dGNG)
}
