% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{design_model}
\alias{design_model}
\title{Title}
\usage{
design_model(
  data,
  design,
  model = NULL,
  prior = NULL,
  add_acc = TRUE,
  rt_resolution = 0.02,
  verbose = TRUE,
  compress = TRUE,
  rt_check = TRUE
)
}
\arguments{
\item{rt_check}{}
}
\description{
Title
}
