% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMC2-package.R
\docType{package}
\name{EMC2-package}
\alias{EMC2}
\alias{EMC2-package}
\title{EMC2: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
