% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{fix_custom_kernel_pointers}
\alias{fix_custom_kernel_pointers}
\title{Reset pointers of custom C++ trend kernels to an emc object}
\usage{
fix_custom_kernel_pointers(emc, pointer_source, model_number = 1)
}
\arguments{
\item{emc}{A target emc object with missing pointers}

\item{pointer_source}{Either an \code{emc2_trend} object with valid pointers, or
an emc object with valid pointers.}

\item{model_number}{If \code{pointer_source} is a raw trend list and \code{emc} is a joint model, specifies
which model the trend belongs to. Defaults to 1.}
}
\value{
An emc object with the custom pointers re-instated.
}
\description{
When an emc object is loaded from disk, or returned by forked processes, the pointers to custom kernels
need to be re-created. This is a convenience function to do this.
}
