% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{fix_custom_kernel_pointers}
\alias{fix_custom_kernel_pointers}
\title{Reset pointers of custom C++ trend kernels to an emc object}
\usage{
fix_custom_kernel_pointers(emc, pointer_source)
}
\arguments{
\item{emc}{A target emc object with missing pointers}

\item{pointer_source}{Either a trend object with correct pointers or another emc object with correct pointers}
}
\value{
An emc object with the custom pointers re-instated.
}
\description{
When an emc object is loaded from disk, or returned by forked processes, the pointers to custom kernels
need to be re-created. This is a convenience function to do this.
}
