% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{gd_summary}
\alias{gd_summary}
\title{gd_summary}
\usage{
gd_summary(samplers, no_print = TRUE, digits = 2)
}
\arguments{
\item{samplers}{Samples object with multiple chains}

\item{no_print}{Boolean for printing}

\item{digits}{}
}
\value{
List of two lists names psrf and mpsrf.
}
\description{
Summarizes gelman_diag statistics for a samplers object, invisibly returning
a list of two lists containing univarite (psrf) and multivariate (mpsrf)
statistics.
}
