% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{get_BayesFactor}
\alias{get_BayesFactor}
\title{Bayes Factors}
\usage{
get_BayesFactor(MLL1, MLL2)
}
\arguments{
\item{MLL1}{Numeric. Marginal likelihood of model 1. Obtained with \code{run_bridge_sampling()}}

\item{MLL2}{Numeric. Marginal likelihood of model 2. Obtained with \code{run_bridge_sampling()}}
}
\value{
The BayesFactor for model 1 over model 2
}
\description{
returns the Bayes Factor for two models
}
\examples{
\donttest{
# Normally one would compare two different models
# Here we use two times the same model:
M1 <- M0 <- run_bridge_sampling(samples_LNR, both_splits = FALSE, cores_for_props = 1)
get_BayesFactor(M1, M0)
}
}
