% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_single.R
\name{get_prior_single}
\alias{get_prior_single}
\title{Prior specification for single subject sampling.}
\usage{
get_prior_single(
  prior = NULL,
  n_pars = NULL,
  sample = T,
  N = 1e+05,
  type = "alpha",
  design = NULL
)
}
\arguments{
\item{prior}{A list of the prior containing the mean (theta_mu_mean) and variance (theta_mu_var).}

\item{n_pars}{Argument used by the sampler, best left NULL. In user case inferred from the design}

\item{sample}{Whether to sample from the prior. Default is TRUE}

\item{N}{How many samples to draw from the prior}

\item{type}{String specifying on what you want the prior. For single subject only "alpha" is a valid option.}

\item{design}{The design obtained from \code{make_design}}
}
\value{
A list of samples from the prior
}
\description{
With this type of sampling, we assume that one, or multiple, subjects are estimated without any hierarchical link.
We need to specify a prior for each parameter. For now, the package assumes a multivariate normal prior on each parameter.
Thus you need to specify prior$theta_mu_mean a vector with an entry for each parameter. Furthermore you need a prior covariance matrix prior$theta_mu_var.
Default is: prior <- list(theta_mu_mean = rep(0, n_pars), theta_mu_var = diag(rep(0, n_pars)))
}
\details{
Specific constraints, such as parameter x needs to be larger than 0, are enforced through transformations performed by the models.
Thus you need to be mindful that even though the prior might be normally distributed, the transformed prior might look very differently.
Set \code{sample = TRUE} and pass your design object to \code{design} to sample from the prior. You can then plot the samples using \code{plot_prior}
}
