% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_standard.R
\name{get_prior_standard}
\alias{get_prior_standard}
\title{Prior specification or prior sampling for standard estimation.}
\usage{
get_prior_standard(
  prior = NULL,
  n_pars = NULL,
  sample = TRUE,
  N = 1e+05,
  type = "mu",
  design = NULL,
  map = TRUE
)
}
\arguments{
\item{prior}{A named list containing the prior mean (theta_mu_mean) and
variance (theta_mu_var). Default prior created if NULL}

\item{n_pars}{Argument used by the sampler, best left NULL. In user case inferred from the design}

\item{sample}{Whether to sample from the prior. Default is TRUE}

\item{N}{How many samples to draw from the prior, default 1e5}

\item{type}{FIX ME}

\item{design}{The design obtained from \code{make_design}, required when map = TRUE}

\item{map}{Boolean, default TRUE reverses malformation used by model to make
sampled parameters unbounded}
}
\value{
A list with a single entry "mu" of samples from the prior (if sample = TRUE) or else a prior object
}
\description{
Prior specification or prior sampling for standard estimation.
}
