% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{get_trend_pnames}
\alias{get_trend_pnames}
\title{Get parameter types from trend object}
\usage{
get_trend_pnames(trend)
}
\arguments{
\item{trend}{A trend object created by make_trend()}
}
\value{
A character vector of parameter names used in the trend
}
\description{
Get parameter types from trend object
}
\examples{
trend <- make_trend(par_names = "v", cov_names = "trial", kernels = "exp_incr")
get_trend_pnames(trend)

}
