% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_SDT.R
\name{hUVSD}
\alias{hUVSD}
\title{Hierarchical Unequal-Variance Signal Detection (hUVSD) Model for Binary Responses}
\usage{
hUVSD()
}
\value{
A model list with all the necessary functions to sample
}
\description{
Binary choice SDT model parameterized with sensitivity (d) and bias (c),
following Lages (2024).
}
\details{
Model parameters are:
d (sensitivity, distance between signal and noise means)
c (bias, deviation from the midpoint between means)
sd (standard deviation of the signal distribution, noise SD fixed at 1)
}
