% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{init_chains}
\alias{init_chains}
\title{Initialize chains}
\usage{
init_chains(
  samplers,
  start_mu = NULL,
  start_var = NULL,
  verbose = FALSE,
  particles = 1000,
  cores_per_chain = 1,
  cores_for_chains = length(samplers)
)
}
\arguments{
\item{samplers}{List of chains made by make_samplers}

\item{start_mu}{Mean of multivariate normal}

\item{start_var}{Variance covariance matrix of multivariate normal}

\item{verbose}{Report progress}

\item{particles}{Number of starting values}

\item{n_cores}{Number of process cores to use}
}
\value{
Samplers object with chains initialized
}
\description{
Adds a set of set of start points to each chain samples from a multivariate
normal
}
