% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LBA.R
\name{LBA}
\alias{LBA}
\title{The Linear Ballistic Accumulator (LBA) model}
\usage{
LBA()
}
\value{
A model list with all the necessary functions to sample
}
\description{
The Linear Ballistic Accumulator, proposes that for each choice alternative, ballistic accumulators race towards a common bound.
The first accumulator to reach the bound determines the choice made. The time taken to reach the threshold determines the response times. For details see \verb{Brown & Heathcote, 2008}
}
\details{
The core parameters of the LBA are the drift rate \code{v}, the response threshold \code{B},
between trial variation in drift rate \code{sv}, between trial variation in startpoint of the drift rate \code{A}, and non-decision time \code{t0}.
Frequently \code{sv} is fixed to 1 to satisfy scaling constraints.

All parameters are estimated on the log scale, except for the drift rate which is estimated on the real line.

Here we use the b = B + A parameterization, which ensures that the response threshold is always higher than the between trial variation in start point of the drift rate.
Also, rates are sampled from normal distributions truncated to be always positive.
}
