% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{make_data}
\alias{make_data}
\title{Simulates data}
\usage{
make_data(
  p_vector,
  design,
  model = NULL,
  trials = NULL,
  data = NULL,
  expand = 1,
  mapped_p = FALSE,
  LT = NULL,
  UT = NULL,
  LC = NULL,
  UC = NULL,
  Fcovariates = NULL,
  return_Ffunctions = FALSE
)
}
\arguments{
\item{p_vector}{parameter vector}

\item{design}{design created by make_Design}

\item{model}{usually an attribute of design but if not must be given here.}

\item{trials}{number of trials per design cell}

\item{data}{A supplied that determines the design, with data sorted by subjects
and a trials = 1:n trials/subject factor added (or overwriting any existing)}

\item{expand}{replicates the design expand times}

\item{mapped_p}{if true instead returns a data frame with one row per design
cell and columns for each parameter specifying how they are mapped to the
design cells.}

\item{LT}{lower truncation bound below which data are removed}

\item{UT}{upper truncation bound above which data are removed}

\item{LC}{lower censoring bound below which data are turned into NA}

\item{UC}{upper censoring bound above which data are turned into NA}

\item{Fcovariates}{either a data frame of covariate valeus with the same
number of rows as the data or a list of functions specifying covariates for
each trial. Must have names specified in the design Fcovariates argument.}

\item{return_Ffunctions}{if false covariates are not returned}
}
\description{
Simulates data based on design and parameter vector arguments using the
random function for a model (usually a design attribute, but if not can be
supplied separately) by one of two methods
\enumerate{
\item creating a fully crossed and balanced design specified by the design
argument's formula list (Flist) and factor contrast list (Clist, if it is
null the data frame creation defaults are used) with number of trials per
cell specified by the trials argument ... or if the data argument is non-null
\item using the design implicit in a data frame supplied, which allows creation
}
}
\section{of unbalanced and other irregular designs, and replacing previous data with}{
simulated data.
Accuracy scored by design matchfun argument.
}

