% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{Simulates data}
\usage{
make_data(
  p_vector,
  design,
  model = NULL,
  trials = NULL,
  data = NULL,
  expand = 1,
  mapped_p = FALSE,
  Fcovariates = NULL,
  return_Ffunctions = FALSE,
  LT = 0,
  UT = Inf,
  LC = 0,
  UC = Inf,
  LCresponse = TRUE,
  UCresponse = TRUE,
  LCdirection = TRUE,
  UCdirection = TRUE,
  force_direction = FALSE,
  force_response = FALSE,
  rtContaminantNA = FALSE
)
}
\arguments{
\item{p_vector}{parameter vector used to simulate data, or a matrix with one
row per subject (with corresponding row names) or a lsit of tables produced
by plot_density (in which case posterior medians are used to simulate data)}

\item{design}{design created by make_Design}

\item{model}{usually an attribute of design but if not must be given here.}

\item{trials}{number of trials per design cell}

\item{data}{If supplied that determines the design, with data sorted by subjects
and a trials = 1:n trials/subject factor added (or overwriting any existing).
Truncation and censoring information also taken from data attributes and presence
of 0/Inf and NA in rt column and NA in R column.}

\item{expand}{replicates the design expand times}

\item{mapped_p}{if true instead returns a data frame with one row per design
cell and columns for each parameter specifying how they are mapped to the
design cells.}

\item{Fcovariates}{either a data frame of covariate values with the same
number of rows as the data or a list of functions specifying covariates for
each trial. Must have names specified in the design Fcovariates argument.}

\item{return_Ffunctions}{if false covariates are not returned}

\item{LT}{lower truncation bound below which data are removed (scalar or subject named vector)}

\item{UT}{upper truncation bound above which data are removed (scalar or subject named vector)}

\item{LC}{lower censoring bound (scalar or subject named vector)}

\item{UC}{upper censoring bound (scalar or subject named vector)}

\item{LCresponse}{Boolean, default TRUE, if false set LC response to NA}

\item{UCresponse}{Boolean, default TRUE, if false set UC response to NA}

\item{LCdirection}{Boolean, default TRUE, set LC rt to 0, else to NA}

\item{UCdirection}{Boolean, default TRUE, set LC rt to Inf, else to NA}

\item{force_direction}{Boolean, take direction from argument not data (default FALSE)}

\item{force_response}{Boolean, take response from argument not data (default FALSE)}

\item{rtContaminantNA}{Boolean, TRUE sets contaminant trial rt to NA, if FALSE
(the default) direction is taken from data or LCdirection or UCdirection (NB
if both of these are false an error occurs as then contamination is not identifiable).}
}
\value{
None
}
\description{
Simulates data based on design and parameter vector arguments using the
random function for a model (usually a design attribute, but if not can be
supplied separately) by one of two methods
\enumerate{
\item creating a fully crossed and balanced design specified by the design
argument's formula list (formuula) and factor contrast list (contrasts), if it is
null the data frame creation defaults are used) with number of trials per
cell specified by the trials argument ... or if the data argument is non-null
\item using the design implicit in a data frame supplied, which allows creation
of unbalanced and other irregular designs, and replacing previous data with
simulated data.
Accuracy scored by design matchfun argument.
}
}
