% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{make_data}
\alias{make_data}
\title{Title}
\usage{
make_data(
  p_vector,
  design,
  model = NULL,
  trials = NULL,
  data = NULL,
  expand = 1,
  mapped_p = FALSE,
  LT = NULL,
  UT = NULL,
  LC = NULL,
  UC = NULL,
  Fcovariates = NULL,
  n_cores = 1,
  return_Ffunctions = FALSE
)
}
\arguments{
\item{return_Ffunctions}{}
}
\description{
Title
}
