% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{make_design}
\alias{make_design}
\title{make_design()}
\usage{
make_design(
  Flist,
  Ffactors,
  Rlevels,
  model,
  Clist = NULL,
  matchfun = NULL,
  constants = NULL,
  Fcovariates = NULL,
  Ffunctions = NULL,
  adapt = NULL,
  report_p_vector = TRUE,
  Rmap = NULL
)
}
\arguments{
\item{Flist}{Formula list. Linear models for each parameter type (e.g., y ~ x)}

\item{Ffactors}{Formula factors. All factors used in design formulas.}

\item{Rmap}{}
}
\description{
This function returns a list with the model design.
}
