% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_missing}
\alias{make_missing}
\title{make_missing}
\usage{
make_missing(
  data,
  LT = 0,
  UT = Inf,
  LC = 0,
  UC = Inf,
  LCresponse = TRUE,
  UCresponse = TRUE,
  LCdirection = TRUE,
  UCdirection = TRUE
)
}
\arguments{
\item{data}{Data frame with rt and R columns}

\item{LT}{lower truncation bound below which data are removed (scalar or subject named vector)}

\item{UT}{upper truncation bound above which data are removed (scalar or subject named vector)}

\item{LC}{lower censoring bound (scalar or subject named vector)}

\item{UC}{upper censoring bound (scalar or subject named vector)}

\item{LCresponse}{Boolean, default TRUE, if false set LC response to NA}

\item{UCresponse}{Boolean, default TRUE, if false set UC response to NA}

\item{LCdirection}{Boolean, default TRUE, set LC rt to 0, else to NA}

\item{UCdirection}{Boolean, default TRUE, set LC rt to Inf, else to NA}
}
\value{
Truncated and censored data frame
}
\description{
Truncate or censor data. is.na(rt) not truncated or censored.
}
