% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{make_prior}
\alias{make_prior}
\title{make_prior}
\usage{
make_prior(
  design,
  pmean = NULL,
  psd = NULL,
  update = NULL,
  verbose = TRUE,
  update_print = TRUE
)
}
\arguments{
\item{design}{Design for which a prior is constructed}

\item{pmean}{Named vector of prior means, or an unnamed scalar, which is then used for all.}

\item{psd}{Named vector of prior standard deviations, , or an unnamed scalar, which is then used for all.}

\item{update}{Another prior from which to copy meand and sds.}

\item{verbose}{Boolean (default true) print values of prior to console (if
update only new values).}

\item{update_print}{When verbose print only new values (default TRUE)}
}
\value{
An EMC prior object
}
\description{
Makes priors for single and standard model (in the assuming no correlation)
based an a design, by providing prior means (pmean) and standard deviations
(psd), and possibly these values taken from another prior (supplied by the
update argument), as long as those values have the same name as in the current
design and are not specified in pmean and psd. Where a value is not supplied
in arguments the user is prompted to enter numeric values (or functions that
evaluate to numbers).
}
