% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{make_prior}
\alias{make_prior}
\title{make_prior}
\usage{
make_prior(
  design,
  pmean = NULL,
  psd = NULL,
  update = NULL,
  verbose = TRUE,
  update_print = TRUE,
  type = "standard",
  pscale = NULL,
  df = NULL
)
}
\arguments{
\item{design}{Design for which a prior is constructed}

\item{pmean}{Named vector of prior means, or an unnamed scalar, which is then used for all.}

\item{psd}{Named vector of prior standard deviations, or an unnamed scalar, which is then used for all. If NA, will assume 1 for all parameters}

\item{update}{Another prior from which to copy values}

\item{verbose}{Boolean (default true) print values of prior to console (if
update only new values).}

\item{update_print}{When verbose print only new values (default TRUE)}

\item{type}{What type of model you plan on using, choice of standard, diagonal and single for this function}

\item{pscale}{For hierarchical models, the prior on the scale of the variances. If NA, will assume 1 for all parameters}

\item{df}{For hierarchical models, the prior on the degrees of freedom of the variances. If NA will assume 2 for all parameters}
}
\value{
An EMC prior object
}
\description{
Makes priors based on a design.
For hierarchical models you specify prior on mean of the group mean (pmean) and standard deviations
(psd), and possibly these values taken from another prior (supplied by the
update argument), as long as those values have the same name as in the current
design and are not specified in pmean and psd. Where a value is not supplied
in arguments the user is prompted to enter numeric values (or functions that
evaluate to numbers).
}
