% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{make_random_effects}
\alias{make_random_effects}
\title{Make random effects.}
\usage{
make_random_effects(
  design,
  group_means,
  n_subj,
  variance_proportion = 0.2,
  covariances = NULL
)
}
\arguments{
\item{design}{A design list. The design as specified by \code{make_design}}

\item{group_means}{A numeric vector. The group level means for each parameter, in the same order as \code{sampled_p_vector(design)}}

\item{n_subj}{An integer. The number of random effects to create.}

\item{variance_proportion}{A double. Optional. If covariances aren't specified, the variances will be created by multiplying the means by this number. The covariances will be 0.}

\item{covariances}{A covariance matrix. Optional. Specify the intended covariance matrix.}
}
\value{
A matrix of random effects.
}
\description{
Make random effects.
}
