% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{make_samplers}
\alias{make_samplers}
\title{Creates a sampler from data and design combined.}
\usage{
make_samplers(
  data_list,
  design_list,
  model_list = NULL,
  type = c("standard", "diagonal", "blocked", "factor", "single")[1],
  n_chains = 3,
  rt_resolution = 0.02,
  prior_list = NULL,
  par_groups = NULL,
  n_factors = NULL,
  constraintMat = NULL
)
}
\arguments{
\item{data_list}{A dataframe of data, or a list of a dataframe. Should have the column subjects as identifiers.}

\item{design_list}{A list with a prespecified design made with make_design.}

\item{model_list}{A model list, if empty will use the model specified in the design_list.}

\item{type}{A string indicating whether to run a standard group-level, or blocked, diagonal, factor, or single.}

\item{n_chains}{An integer. Specifies the amount of mcmc chains to be run. Should be more than 1 to get gelman diagnostics.}

\item{rt_resolution}{A double. Used for compression, rts will be binned based on this resolution.}

\item{prior_list}{A list of priors for the group level. Prior distributions should match the type argument.}

\item{par_groups}{A vector. Only to be specified with type blocked \code{c(1,1,1,2,2)} means first three parameters first block, last two parameters in the second block}

\item{n_factors}{An integer. Only to be specified with type factor.}

\item{constraintMat}{A matrix of rows equal to the number of estimated parameters, and columns equal to the number of factors, only to be specified with type factor.
If null will use default settings as specified in Innes et al. 2022}
}
\value{
a list of samplers
}
\description{
This function is used to initialize samplers using the data and the prespecified design.
}
