% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{merge_samples}
\alias{merge_samples}
\title{Merge samples}
\usage{
merge_samples(samplers)
}
\arguments{
\item{samplers}{A samplers object, commonly the output of \code{run_emc()}}
}
\value{
An unlisted samplers object with all chains merged
}
\description{
Merges samples from all chains as one unlisted object.
}
\details{
Note that all sampling stages are included in the merged output,
including iterations from the \code{preburn}, \code{burn}, and \code{adapt} stages.
\code{merge_samples(samplers)$samples$stage} shows the corresponding sampling stages.
}
