% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LBA.R
\name{MlbaB}
\alias{MlbaB}
\title{lba_B Model accommodating missing values (truncation and censoring)}
\usage{
MlbaB()
}
\value{
A model list with all the necessary functions to sample
}
\description{
lba_B Model accommodating missing values (truncation and censoring)
}
