% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.r
\name{n_eff}
\alias{n_eff}
\title{Effective sample size (ESS)}
\usage{
n_eff(mcmc_list, version = c("old", "new"))
}
\arguments{
\item{mcmc_list}{A \code{coda::mcmc.list} object containing MCMC draws.}

\item{version}{Character string, either \code{"old"} or \code{"new"}. \code{"old"} (default)
calls \code{\link[coda:effectiveSize]{coda::effectiveSize()}}, while \code{"new"} uses a vendored implementation
of \code{\link[posterior:ess_basic]{posterior::ess_basic()}}.}
}
\value{
A named numeric vector of effective sample sizes, with names
corresponding to parameters.
}
\description{
Computes the effective sample size (ESS) for each parameter from an MCMC run.
Supports both the legacy \code{\link[coda:effectiveSize]{coda::effectiveSize()}} implementation and the
improved version based on Vehtari et al. (2021).
}
\details{
The \code{"new"} method is adapted directly from the \pkg{posterior}
package's implementation, but included here to avoid introducing additional
package dependencies. The substantive computation is identical, though some
internal details have been simplified.
}
\references{
Vehtari, A., Gelman, A., Simpson, D., Carpenter, B., & Bürkner, P.-C. (2021).
Rank-normalization, folding, and localization: An improved \eqn{\hat{R}} for
assessing convergence of MCMC. \emph{Bayesian Analysis}, \emph{16}(2), 667–718.
}
\keyword{internal}
