% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{p_test}
\alias{p_test}
\title{Posterior parameter tests}
\usage{
p_test(
  x,
  x_name = NULL,
  x_fun = NULL,
  x_fun_name = "fun",
  y = NULL,
  y_name = NULL,
  y_fun = NULL,
  y_fun_name = "fun",
  mapped = FALSE,
  x_subject = NULL,
  y_subject = NULL,
  mu = 0,
  alternative = c("less", "greater")[1],
  probs = c(0.025, 0.5, 0.975),
  digits = 2,
  p_digits = 3,
  print_table = TRUE,
  filter = "sample",
  selection = "mu",
  subfilter = 0
)
}
\arguments{
\item{x}{An EMC2 samplers object}

\item{x_name}{A character string. Name of the parameter to be tested for \code{x}}

\item{x_fun}{Function applied to the MCMC chains to create
variable to be tested.}

\item{x_fun_name}{Name to give to quantity calculated by \code{x_fun}}

\item{y}{A second EMC2 samplers object}

\item{y_name}{A character string. Name of the parameter to be tested for \code{y}}

\item{y_fun}{Function applied to the MCMC chains to create
variable to be tested.}

\item{y_fun_name}{Name to give to quantity calculated by \code{y_fun}}

\item{mapped}{Boolean. Should the samples be mapped back to the design before doing the test?}

\item{x_subject}{Integer or name selecting a subject}

\item{y_subject}{Integer or name selecting a subject}

\item{mu}{Numeric. \code{NULL} value for single sample test if \code{y} is not supplied (default 0)}

\item{alternative}{\code{less} or \code{greater} determining direction of test probability}

\item{probs}{Vector defining quantiles to return.}

\item{digits}{Integer, significant digits for estimates in printed results}

\item{p_digits}{Integer, significant digits for probability in printed results}

\item{print_table}{Boolean (defaults to \code{TRUE}) for printing results table}

\item{filter}{A character string. Specifies from which stage the samples should be taken.}

\item{selection}{A character string designating parameter type (e.g. \code{alpha} or \code{covariance})}

\item{subfilter}{An integer or vector. If it's an integer, iterations up until the value set by \code{subfilter} will be excluded.
If a vector is supplied, only the iterations in the vector will be considered.}
}
\value{
Invisible results table with no rounding.
}
\description{
Modeled after \code{t.test}, returns the credible interval of the parameter or test
and what proportion of the posterior distribution (or the difference in posterior distributions
in case of a two sample test) overlaps with mu.
For a one sample test provide \code{x} and for two sample also provide \code{y}.
Note that fo comparisons within one model, we recommend using \code{savage_dickey()} if the priors
were well chosen.
}
\examples{
\dontrun{
# Run a credible interval test (Bayesian ''t-test'')
p_test(samplers, x_name = "v")
# We can also compare between two sets of samplers
p_test(x = samplers1, x_name = "v", y = samplers2, y_name = "v")

# Or provide custom functions
p_test(x = samplers, x_fun = function(d) d["B_Eaccuracy"] - d["B_Eneutral"])
}
}
