% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pairs_posterior}
\alias{pairs_posterior}
\title{Title}
\usage{
pairs_posterior(
  samples,
  filter = "sample",
  thin = 1,
  subfilter = 0,
  mapped = FALSE,
  selection = c("alpha", "mu", "variance", "covariance", "correlation")[1],
  stat = NULL,
  scale.subjects = TRUE,
  use = NA,
  do_plot = TRUE,
  maxp = 500
)
}
\arguments{
\item{maxp}{}
}
\description{
Title
}
