% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_funcs.R
\name{plot.emc}
\alias{plot.emc}
\title{Plot function for EMC2 samplers objects}
\usage{
\method{plot}{emc}(
  x,
  filter = "sample",
  selection = c("mu", "variance", "correlation"),
  subfilter = 0,
  subject = NULL,
  thin = 1,
  ...
)
}
\arguments{
\item{x}{An object of class \code{emc}}

\item{filter}{A character string indicating the sampling stage to be summarized.
Can be \code{preburn}, \code{burn}, \code{adapt}, or \code{sample}.}

\item{selection}{A character string indicating the parameter group.
Defaults to \code{mu}, \code{variance}, and \code{correlation}. Other options are \code{covariance}, and \code{alpha}
(i.e., individual-level parameters).}

\item{subfilter}{An integer or vector, defaults to 0. See below for more details.}

\item{subject}{An integer or character vector. Plot samples selected subjects.}

\item{thin}{An integer, only consider samples that are a multiple of \code{thin}.}

\item{...}{Further optional arguments.}
}
\description{
Plots posterior densities and traceplots for EMC2 model parameters.
}
\details{
Note that if \code{subfilter} is set to an integer other than 0 (i.e., the default), all iterations
up to that integer are ignored. If a vector is supplied, only those iterations
are considered.
}
