% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_acfs}
\alias{plot_acfs}
\title{Calls plot_chains to plot auto-correlation functions for each parameter}
\usage{
plot_acfs(
  samples,
  layout = NULL,
  subject = 1,
  selection = "alpha",
  filter = "sample",
  subfilter = 0
)
}
\arguments{
\item{samples}{Single pmwgs object (in which case ACF for one chain is plotted)
or list of pmwgs objects (in which case ACFs for each chain plotted)}

\item{layout}{A 2-vector specifying the layout handled as in plot_chains}

\item{subject}{integer or character vector, if selection = "alpha" picks out subjects(s)}

\item{selection}{String designating parameter type (mu, variance, correlation, alpha = default)}

\item{filter}{which stage to plot (default "sample")}

\item{subfilter}{an integer or vector. If integer it will exclude up until that integer.
If vector it will include everything in that range.}
}
\value{
None
}
\description{
Calls plot_chains to plot auto-correlation functions for each parameter
}
