% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_alpha_recovery}
\alias{plot_alpha_recovery}
\title{Plots recovery of subject-level parameters}
\usage{
plot_alpha_recovery(
  tabs,
  layout = c(2, 3),
  do_ci = TRUE,
  ci_col = "grey",
  cap = 0.05,
  do_rmse = FALSE,
  r_pos = "topleft",
  rmse_digits = 3,
  pearson_digits = 2,
  do_coverage = TRUE,
  coverage_pos = "bottomright",
  coverage_digits = 1,
  spearman_digits = 2
)
}
\arguments{
\item{tabs}{Tables of actual and estimated alpha parameters (with CIs) from plot_pars}

\item{layout}{A vector specifying the layout as in \code{par(mfrow = layout)}}

\item{do_ci}{Boolean (Default TRUE). Add credible intervals to plot?}

\item{ci_col}{Color of CI.}

\item{cap}{Width of CI cap (passed to arrows)}

\item{do_rmse}{Boolean. If \code{TRUE} will add root-mean squared error to plot.
If \code{FALSE} will instead plot the Pearson correlation}

\item{r_pos}{Position of Pearson/RMSE (passed to legend)}

\item{rmse_digits}{Digits for RMSE}

\item{pearson_digits}{Digits for Pearson correlation}

\item{do_coverage}{Boolean (default TRUE) add coverage percentage estimate (otherwise
add Spearman correlation)}

\item{coverage_pos}{Position of Coverage/Pearson (passed to legend)}

\item{coverage_digits}{Digits for coverage}

\item{spearman_digits}{Digits for Spearman correlation}
}
\value{
Invisible list with RMSE, coverage and Pearson and Spearman correlations.
}
\description{
Uses output from \code{plot_pars} to plot true vs. estimated (median with CI)
parameters for each subject.
}
