% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_alpha_recovery}
\alias{plot_alpha_recovery}
\title{Uses output from plot_density to plot true vs. estimated (median with CI) alpha
parameters for each subject (plot density must be called with true values passed
through the pars argument).}
\usage{
plot_alpha_recovery(
  tabs,
  layout = c(2, 3),
  do_ci = TRUE,
  ci_col = "grey",
  cap = 0.05,
  do_rmse = FALSE,
  r_pos = "topleft",
  rmse_digits = 3,
  pearson_digits = 2,
  do_coverage = FALSE,
  coverage_pos = "bottomright",
  coverage_digits = 1,
  spearman_digits = 2
)
}
\arguments{
\item{tabs}{Tables of actual and estimated alpha parameters (with CIs) from plot_density}

\item{layout}{par(mfrow=layout) for parameters}

\item{do_ci}{Boolean (Defualt TRUE). Add CIs to plot?}

\item{ci_col}{Color of CI.}

\item{cap}{Width of CI cap (passed to arrows)}

\item{do_rmse}{Boolean (defualt FALSE) Add root-mean squared error to plot
instead of default Pearson correlation}

\item{r_pos}{Position of Pearson/RMSE (passed to legend)}

\item{rmse_digits}{Digits for RMSE}

\item{pearson_digits}{Digits for Pearson correlation}

\item{do_coverage}{Boolean (defualt TRUE) add coverage percentage estimate (otherwise
add Spearman correlation)}

\item{coverage_pos}{Position of Coverage/Pearson (passed to legend)}

\item{coverage_digits}{Digits for coverage}

\item{spearman_digits}{Digits for Spearman correlation}
}
\value{
List with RMSE, coverage and Pearson and Spearman correlations.
}
\description{
Uses output from plot_density to plot true vs. estimated (median with CI) alpha
parameters for each subject (plot density must be called with true values passed
through the pars argument).
}
