% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_cdf}
\alias{plot_cdf}
\title{Plot defective cumulative distribution functions}
\usage{
plot_cdf(
  input,
  subject = NULL,
  factors = NULL,
  defective_factor = "R",
  layout = NA,
  post_predict = NULL,
  quants = c(0.025, 0.975),
  n_cores = 1,
  n_post = 100,
  predict_args = list(),
  ...
)
}
\arguments{
\item{subject}{An integer or character string selecting a subject from the data.
If specified, only that subject is plotted. By default (NULL), all subjects are plotted.}

\item{factors}{A character vector of the factor names in the design to aggregate across.
Defaults to all factors (i.e., NULL).}

\item{defective_factor}{A character string specifying the factor to compute defective CDFs for. Defaults to "R".}

\item{layout}{A vector indicating which layout to use as in par(mfrow = layout). If NA, will automatically generate an appropriate layout.}

\item{post_predict}{A data frame containing posterior predictive datasets. Should have the same columns as data, with an additional postn column indicating the index of the predicted dataset.}

\item{quants}{A numeric vector specifying the quantiles for the credible intervals (defaults to c(0.025, 0.975)).}

\item{n_cores}{An integer specifying the number of cores with which to calculate predictives if specified.}

\item{n_post}{An optional integer specifying how many posterior predictives to generate if the input is an emc object.}

\item{...}{Optional arguments that can be passed to plotting functions for the real data.}

\item{input.}{Either an emc object or a data frame. The data frame must be in EMC2 format with at least subjects, rt, and the defective factor columns.}

\item{predicted_plot_args}{A list of plotting arguments to be used for the predicted data (e.g., col, lty).}
}
\value{
Invisibly returns a data frame containing RT quantiles and mean proportions per factor combination and defective factor level.
If post_predict is provided or the input is an emc object, predicted values are included.
}
\description{
Plots panels that contain a set of cumulative distribution functions (CDFs) for each level of the specified defective factor in the data.
These CDFs are defective; their maximum values are relative to the respective
proportions of the defective factor levels. Across all levels, the maximum value sums to 1.
Optionally, posterior predictive CDFs can be overlaid.
}
\examples{
# Plot defective cdf for each subject and the factor combination in the design:
plot_cdf(forstmann)
# or for one subject:
plot_cdf(forstmann, subject = 1)
# Now collapsing across subjects and using a different defective factor:
plot_cdf(forstmann, factors = "S", defective_factor = "E")
# Or plot posterior predictives with an emc object
plot_cdf(samples_LNR, n_post = 10)
}
