% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_chains}
\alias{plot_chains}
\title{Plot MCMC chains}
\usage{
plot_chains(
  pmwg_mcmc,
  layout = NA,
  subject = NA,
  ylim = NULL,
  selection = "alpha",
  filter = "sample",
  thin = 1,
  subfilter = 0,
  plot_acf = FALSE,
  acf_chain = 1
)
}
\arguments{
\item{pmwg_mcmc}{A list of samplers or samplers converted to mcmc objects.}

\item{layout}{A vector or matrix specifying the layout as in par(mfrow = layout).}

\item{subject}{Integer or character vector, if selection = "alpha" picks out subject(s)}

\item{ylim}{The y limits of the chain plot.}

\item{selection}{A string. Specifies which chains you want to plot.}

\item{filter}{A string. Specifies which stage you want to plot.}

\item{thin}{An integer. Specify if you want to thin the chains by a factor n.}

\item{subfilter}{An integer or vector. If integer it will exclude up until
that integer. If vector it will include everything in that range.}

\item{plot_acf}{Boolean (default FALSE). If TRUE will also plot autocorrelation
for the chain specified in acf_chain.}

\item{acf_chain}{Integer. For which chain to plot the acf, if plot_acf = TRUE.}
}
\description{
Plot MCMC chains
}
