% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_chains}
\alias{plot_chains}
\title{Plot MCMC chains}
\usage{
plot_chains(
  samplers,
  layout = NA,
  subject = NA,
  ylim = NULL,
  selection = "mu",
  filter = "sample",
  thin = 1,
  subfilter = 0,
  plot_acf = FALSE
)
}
\arguments{
\item{samplers}{A list of samplers.}

\item{layout}{A vector specifying the layout as in par(mfrow = layout).
If NA (default) will use CODA defaults (unless \code{plot_acf = TRUE}), if NULL use current.}

\item{subject}{Integer or character vector. Only applicable if selection = "alpha". Will plot only these subject(s).
NA (default) will plot all.}

\item{ylim}{A vector. The y limits of the chain plot.}

\item{selection}{String. Which parameter type to plot ("alpha", "mu", "variance", "covariance", "correlation").
"LL" will plot the log-likelihood chains.}

\item{filter}{A string. Specifies from which stage you want to plot the MCMC chains ("preburn", "burn", "adapt", "sample")                        .}

\item{thin}{An integer. Will keep only iterations of the MCMC chains that are a multiple of \code{thin}.}

\item{subfilter}{An integer or vector. If integer it will exclude up until
that integer. If vector it will include everything in that range.}

\item{plot_acf}{Boolean. If TRUE will plot autocorrelation
for the first MCMC chain (of the default three chains).}
}
\value{
None
}
\description{
Plots the trace plots of the MCMC chains on top of each other to visualize convergence
and chain stability.
}
\examples{
\dontrun{
# For a set of samplers run using run_emc:
plot_chains(samplers)
# Or for the second subject:
plot_chains(samplers, subject = 2)

# Can also plot the autocorrelation of for example the group-level mean:
plot_chains(samplers, selection = "mu", plot_acf = TRUE)
}
}
