% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_chains}
\alias{plot_chains}
\title{Title Plots chains for a variety of quantities produced by sampling}
\usage{
plot_chains(
  pmwg_mcmc,
  layout = NA,
  subject = NA,
  ylim = NULL,
  selection = "alpha",
  filter = "sample",
  thin = 1,
  subfilter = 0,
  plot_acf = FALSE,
  acf_chain = 1,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
Title Plots chains for a variety of quantities produced by sampling
}
