% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_density}
\alias{plot_density}
\title{Title}
\usage{
plot_density(
  pmwg_mcmc,
  layout = c(2, 3),
  selection = "alpha",
  filter = "sample",
  thin = 1,
  subfilter = 0,
  mapped = FALSE,
  plot_prior = TRUE,
  n_prior = 1000,
  xlim = NULL,
  ylim = NULL,
  show_chains = FALSE,
  do_plot = TRUE,
  subject = NA,
  add_means = FALSE,
  pars = NULL,
  probs = c(0.025, 0.5, 0.975),
  bw = "nrd0",
  adjust = 1
)
}
\arguments{
\item{adjust}{}
}
\description{
Title
}
