% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_fit}
\alias{plot_fit}
\title{Title}
\usage{
plot_fit(
  data,
  pp,
  subject = NULL,
  factors = NULL,
  stat = NULL,
  stat_name = "",
  ci = c(0.025, 0.5, 0.975),
  do_plot = TRUE,
  xlim = NULL,
  ylim = NULL,
  layout = NULL,
  mfcol = TRUE,
  probs = c(1:99)/100,
  data_lwd = 2,
  fit_lwd = 1,
  qp_cex = 1,
  q_points = c(0.1, 0.3, 0.5, 0.7, 0.9),
  pqp_cex = 0.5,
  lpos = "topleft",
  signalFactor = "S",
  zROC = FALSE,
  qfun = NULL,
  lim = NULL,
  rocfit_cex = 0.5
)
}
\arguments{
\item{rocfit_cex}{}
}
\description{
Title
}
