% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_fit_choice}
\alias{plot_fit_choice}
\title{Plots choice data with no response times}
\usage{
plot_fit_choice(
  data,
  pp,
  subject = NULL,
  factors = NULL,
  functions = NULL,
  stat = NULL,
  stat_name = "",
  adjust = 1,
  ci = c(0.025, 0.5, 0.975),
  do_plot = TRUE,
  xlim = NULL,
  ylim = NULL,
  main = "",
  layout = NULL,
  mfcol = TRUE,
  signalFactor = "S",
  zROC = FALSE,
  qfun = qnorm,
  lim = NULL,
  rocfit_cex = 0.5
)
}
\arguments{
\item{data}{Data frame with subjects and R factors, and possibly other factors
and an rt column}

\item{pp}{Posterior predictives created by post_predict}

\item{subject}{Integer or string picking out subject(s).}

\item{factors}{Character vector of factors in data to display separately. If
NULL (default) use names of all columns in data except "trials","R", and "rt".
Omitted factors are aggregated over. If NA treats entire data set as a single cell.
If stat is used default is changed to NA.}

\item{functions}{A named list of functions that create new factors which cna then be
used by the factors and stat arguments.}

\item{stat}{A function that takes a the data and returns a single value.}

\item{stat_name}{A string naming what the stat argument calculates.}

\item{adjust}{Control of smoothing in density plots}

\item{ci}{Credible interval and central tendency quantiles for return when
stat argument is supplied (default c(.025,.5,.975))}

\item{do_plot}{Boolean (default TRUE) for making a plot}

\item{xlim}{x-axis plot limit, 2-vector (same for all) or matrix (one row for each paramter)}

\item{ylim}{y-axis plot limit, 2-vector (same for all) or matrix (one row for each paramter)}

\item{main}{Text title, pasted before cell name.}

\item{layout}{2-vector specifying par(mfrow) or par(mfcol) (default NULL use current,
NA keeps par currently active).}

\item{mfcol}{Boolean, default TRUE use mfcol else mfrow.}

\item{signalFactor}{Character name of factor for the signal}

\item{zROC}{Logical, plot Z transformed ROC (default FALSE )}

\item{qfun}{Type of "Z" transform (default probit)}

\item{lim}{x = y limit for ROC plots}

\item{rocfit_cex}{Size of points in ROC plot (default 0.5)}
}
\value{
If stat argument is provided a matrix of observed values and predicted quantiles
}
\description{
Plots choice data with no response times
}
