% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pars}
\alias{plot_pars}
\title{Plots density for parameters.}
\usage{
plot_pars(
  samplers,
  layout = c(2, 3),
  use_par = NULL,
  selection = "mu",
  filter = "sample",
  thin = 1,
  subfilter = 0,
  mapped = FALSE,
  plot_prior = TRUE,
  xlim = NULL,
  ylim = NULL,
  prior_xlim = c(0.1, 0.9),
  show_chains = FALSE,
  do_plot = TRUE,
  subject = NA,
  add_means = FALSE,
  pars = NULL,
  probs = c(0.025, 0.5, 0.975),
  bw = "nrd0",
  adjust = 1,
  lpos = "topright",
  digits = 3
)
}
\arguments{
\item{samplers}{A list of samplers.}

\item{layout}{A vector specifying the layout as in par(mfrow = layout).
If NA or NULL use current plot layout.}

\item{use_par}{Character vector of names of parameters to plot (default NULL = plot all)}

\item{selection}{A string. Which parameter type to plot ("alpha", "mu", "variance", "covariance", "correlation").}

\item{filter}{A string. Specifies from which stage you want to plot the densities ("preburn", "burn", "adapt", "sample")}

\item{thin}{An integer. Will keep only iterations of the MCMC chains that are a multiple of \code{thin}.}

\item{subfilter}{An integer or vector. If integer it will exclude up until
that integer. If vector it will include everything in that range.}

\item{mapped}{Boolean. If TRUE plots the parameters mapped back to experimental design
otherwise plots the sampled parameters.}

\item{plot_prior}{Boolean. If \code{TRUE} will overlay prior density in the plot (in red)}

\item{xlim}{x-axis plot limit. If a vector is supplied will use the same axes for all.}

\item{ylim}{y-axis plot limit. If a vector is supplied will use the same axes for all.}

\item{prior_xlim}{A vector giving upper and lower quantiles of prior when choosing
xlim if \code{plot_prior = TRUE}. If set to NULL xlim is used instead.}

\item{show_chains}{Boolean (default FALSE) plots a separate density for each chain.}

\item{do_plot}{Boolean. Set to \code{FALSE} to only return the parameter credible intervals and omit the plots.}

\item{subject}{Integer or character vector. Only applicable if \code{selection = "alpha"}. Will plot only these subject(s).
NA (default) will plot all.}

\item{add_means}{Boolean. Whether to add parameter means as an attribute
to return invisibly}

\item{pars}{Named vector or matrix of known/simulated parameters, or the output of
\code{plot_pars}, in which case the posterior medians are extracted. If xlim is not supplied,}

\item{probs}{Vector. The quantiles of the selected parameters to return invisibly.}

\item{bw}{number or string bandwidth for density (default "nrd0"). See \code{?density}.}

\item{adjust}{Numeric. Adjustment for density plot. See \code{?density}}

\item{lpos}{Character. Position of the contraction in the plot}

\item{digits}{Integer. Rounding of contraction}
}
\value{
invisibly return quantiles for the selected parameters
}
\description{
Plots the posterior and prior density for selected parameters of a model.
}
\examples{
\dontrun{
# For a set of samplers plot prior and posterior densities:
plot_pars(samplers)
# Or just get the quantiles and omit the plot
quants <- plot_pars(samplers, selection = "variance", do_plot = FALSE)
print(quants)
}
}
