% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pars}
\alias{plot_pars}
\title{Plots density for parameter estimates.}
\usage{
plot_pars(
  pmwg_mcmc,
  layout = c(2, 3),
  use_par = NULL,
  selection = "alpha",
  filter = "sample",
  thin = 1,
  subfilter = 0,
  mapped = FALSE,
  plot_prior = TRUE,
  n_prior = 1000,
  xlim = NULL,
  ylim = NULL,
  prior_xlim = NULL,
  show_chains = FALSE,
  do_plot = TRUE,
  subject = NA,
  add_means = FALSE,
  pars = NULL,
  probs = c(0.025, 0.5, 0.975),
  bw = "nrd0",
  adjust = 1,
  do_contraction = TRUE,
  lpos = "topright",
  digits = 3,
  use_main = NULL
)
}
\arguments{
\item{pmwg_mcmc}{A list of samplers or samplers converted to mcmc objects.}

\item{layout}{A 2-vector specifying the layout as in par(mfrow = layout).
If NA or NULL use current.}

\item{use_par}{Character vector of names of parameters to plot (default NULL = plot all)}

\item{selection}{String designating parameter type (mu, variance, correlation, alpha = default)}

\item{filter}{A string. Specifies which stage you want to plot.}

\item{thin}{An integer. Keep only iterations that are a multiple of thin.}

\item{subfilter}{An integer or vector. If integer it will exclude up until
that integer. If vector it will include everything in that range.}

\item{mapped}{Boolean (default FALSE) if TRUE plot parameters mapped to design
otherwise sampled parameters}

\item{plot_prior}{Boolean. Add prior distribution to plot (in red)}

\item{n_prior}{Number of samples to approximate prior (default = 1e3)}

\item{xlim}{x-axis plot limit, 2-vector (same for all) or matrix (one row for each parameter)}

\item{ylim}{y-axis plot limit, 2-vector (same for all) or matrix (one row for each parameter)}

\item{prior_xlim}{A vector giving upper and lower quantiles of prior when choosing xlim.}

\item{show_chains}{Boolean (default FALSE) plot separate density for each chain.}

\item{do_plot}{Boolean (default TRUE) do plot}

\item{subject}{Integer or character vector, if selection = "alpha" picks out
subject(s) (default NA plots all).}

\item{add_means}{Boolean (default FALSE) add parameter means as an attribute
to return}

\item{pars}{Named vector or matrix of true parameters, or the output of
plot_pars, in which case the posterior medians are extracted. If supplied with
no xlim the plot will be adjusted to include the parameter values, which are
plotted as vertical lines.}

\item{probs}{Vector (default c(.025,.5,.975)) for CI and central tendency of return}

\item{bw}{Bandwidth for density plot (see density)}

\item{adjust}{Adjustment for density plot (see density)}

\item{do_contraction}{Print the prior to posterior contraction (1 - var(prior)/var(posterior))}

\item{lpos}{Position of contraction in graph}

\item{digits}{Rounding of contraction}

\item{use_main}{Character string for title to use. If NULL (defualt) auto generated.}
}
\value{
Invisibly returns tables of true and 95\% CIs (for all chains combined
no matter what show_chains is), if do_contraction with a "contraction" attribute.
}
\description{
Plots density for parameter estimates.
}
