% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{plot_prior}
\alias{plot_prior}
\title{plot_prior
Plots prior distributions by simulation}
\usage{
plot_prior(
  prior = NULL,
  design,
  plotp = NULL,
  type = c("standard", "single")[1],
  selection = NULL,
  mapped = TRUE,
  data = NULL,
  N = 1e+05,
  nrep = 10,
  breaks = 50,
  layout = c(3, 3),
  lower = NULL,
  upper = NULL,
  xlim = NA,
  n_factors = NULL,
  ...
)
}
\arguments{
\item{prior}{A list of specifying prior means and covariance, if NULL default
prior (mean=0, variance = 1, uncorrelated)}

\item{design}{Design corresponding to prior}

\item{plotp}{Names of parameters to plot (default NULL plots all)}

\item{type}{Type of prior (standard or single)}

\item{selection}{Select level prior to plot (default "alpha" for single and "mu" for standard)}

\item{mapped}{boolean for mapping mu or alpha back to the design cells on the natural scale.}

\item{data}{data frame, required when mapping involves priors}

\item{N}{Number of prior samples if data not provided}

\item{nrep}{Number of prior samples as multiple of numebr of rows in the data (minus any that violate "ok" constraint)}

\item{breaks}{Histogram breaks parameter}

\item{layout}{par(mfrow) setting (default c(3,3))}

\item{lower}{Lower quantile limit of values plotted. Default NULL all plotted,
numeric same for all parameters, parameter named list parameter specific}

\item{upper}{Upper quantile limit of values plotted. Default NULL all plotted,
numeric same for all parameters, parameter named list parameter specific}

\item{xlim}{List with parameter names of plot x limits or single pair same for all.
Any names not in list or if (default) NA xlim set as min and max.}

\item{n_factors}{Integer. For prior plotting of factor models}
}
\value{
Invisible sampled columns of parameters as a data frame (covariate,
version first column = cell names) or otherwise a matrix
}
\description{
plot_prior
Plots prior distributions by simulation
}
