% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{plot_prior}
\alias{plot_prior}
\title{Convenience function to plot from the prior}
\usage{
plot_prior(
  prior,
  type = NULL,
  add_density = FALSE,
  adjust = 1,
  breaks = 50,
  layout = c(3, 3)
)
}
\arguments{
\item{prior}{A list of prior samples}

\item{type}{Optional. Otherwised inferred from the prior samples}
}
\value{
NULL. Makes a plot of the prior samples
}
\description{
Convenience function to plot from the prior
}
