% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{plot_relations}
\alias{plot_relations}
\title{Plot relations}
\usage{
plot_relations(
  samplers = NULL,
  stage,
  loadings = NULL,
  standardize = T,
  corrs = NULL,
  plot_cred = TRUE,
  plot_means = T,
  do_corr = F,
  only_cred = F,
  nice_names = NULL
)
}
\arguments{
\item{samplers}{A list of samplers}

\item{stage}{Character. The stage from which to take the samples}

\item{loadings}{An array of loadings. Can be alternatively supplied if samplers is not supplied}

\item{standardize}{Boolean. Whether to standardize the loadings only standardizes if loadings isn't supplied.}

\item{corrs}{An array of correlations Can be alternatively supplied if samplers is not supplied}

\item{plot_cred}{Boolean. Whether to plot the credible intervals}

\item{plot_means}{Boolean. Whether to plot the means}

\item{do_corr}{Boolean. If samplers is suplied, whether to take the correlations rather than the loadings}

\item{only_cred}{Boolean. Whether to only plot credible values}

\item{nice_names}{Boolean. Nice names to give the parameters}
}
\description{
An adjusted version of the corrplot package \code{corrplot} tailored to EMC2 and the plotting of factor loadings or correlations.
}
