% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{plot_relations}
\alias{plot_relations}
\title{Plot relations}
\usage{
plot_relations(
  samplers = NULL,
  filter = "sample",
  corrs = NULL,
  plot_cred = TRUE,
  plot_means = T,
  do_corr = F,
  only_cred = F,
  nice_names = NULL,
  ...
)
}
\arguments{
\item{samplers}{A list of samplers}

\item{filter}{Character. The stage from which to take the samples}

\item{corrs}{An array of correlations Can be alternatively supplied if samplers is not supplied}

\item{plot_cred}{Boolean. Whether to plot the credible intervals}

\item{plot_means}{Boolean. Whether to plot the means}

\item{do_corr}{Boolean. If samplers is suplied, whether to take the correlations rather than the loadings}

\item{only_cred}{Boolean. Whether to only plot credible values}

\item{nice_names}{Boolean. Alternative names to give the parameters}

\item{loadings}{An array of loadings. Can be alternatively supplied if samplers is not supplied}

\item{standardize}{Boolean. Whether to standardize the loadings. Only standardizes if loadings isn't supplied.}
}
\description{
An adjusted version of the \code{corrplot} package function \code{corrplot()} tailored
to \code{EMC2} and the plotting of correlations.
}
\examples{
\dontrun{
# For a given set of hierarchical factor model samples we can make a
# loadings matrix plot.
plot_relations(samplers, only_cred = TRUE, plot_cred = TRUE)
# We can also plot the correlations, this works for any model that included correlations
plot_relations(samplers, do_corr = T, plot_means = T, only_cred = T)
}
}
