% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBC.R
\name{plot_sbc_ecdf}
\alias{plot_sbc_ecdf}
\title{Plot the ECDF Difference in SBC Ranks}
\usage{
plot_sbc_ecdf(ranks, layout = NA, add_stats = TRUE)
}
\arguments{
\item{ranks}{A list of named dataframes of the rank statistic}

\item{layout}{Optional. A numeric vector specifying the layout using \code{par(mfrow = layout)}}

\item{add_stats}{Boolean. Should coverage, bias and precision be included in the figure.}
}
\value{
No returns
}
\description{
Plots the difference in observed cumulative rank statistics and the
expected cumulative distribution of a uniform distribution. The blue
shaded areas indicate the 95\% credible interval.
}
