% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBC.R
\name{plot_sbc_ecdf}
\alias{plot_sbc_ecdf}
\title{Plot the ECDF difference in SBC ranks}
\usage{
plot_sbc_ecdf(ranks, layout = NA)
}
\arguments{
\item{ranks}{A list of named dataframes of the rank statistic}

\item{layout}{Optional. A numeric vector specifying the layout using \code{par(mfrow = layout)}}
}
\value{
No returns
}
\description{
Plots the difference in observed cumulative rank statistics and the
expected cumulative distribution of a uniform distribution. The blue
shaded areas indicate the 95\% credible interval.
}
