% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBC.R
\name{plot_sbc_hist}
\alias{plot_sbc_hist}
\title{Plot the Histogram of the Observed Rank Statistics of SBC}
\usage{
plot_sbc_hist(ranks, bins = 10, layout = NA)
}
\arguments{
\item{ranks}{A list of named dataframes of the rank statistic}

\item{bins}{An integer specifying the number of bins to use when plotting the histogram}

\item{layout}{Optional. A numeric vector specifying the layout using \code{par(mfrow = layout)}}
}
\value{
No returns
}
\description{
Note that this plot is dependent on the number of bins, and a more general
visualization is to use \code{plot_sbc_ecdf}
}
