% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_stat}
\alias{plot_stat}
\title{Plot statistics on data}
\usage{
plot_stat(
  input,
  stat_fun,
  stat_name = NULL,
  factors = NULL,
  post_predict = NULL,
  quants = c(0.025, 0.5, 0.975),
  do_plot = TRUE,
  layout = NA,
  n_post = 100,
  n_cores = 1,
  predict_args = list(),
  ...
)
}
\arguments{
\item{stat_fun}{A function that can be applied to the data and returns a single value or a vector of values.}

\item{stat_name}{The name of the calculated quantity}

\item{factors}{A character vector of the factor names in the design to aggregate across.
Defaults to all factors (i.e., NULL).}

\item{post_predict}{A data frame containing posterior predictive datasets. Should have the same columns as data, with an additional postn column indicating the index of the predicted dataset.}

\item{quants}{A numeric vector specifying the quantiles for the credible intervals (defaults to c(0.025, 0.975)).}

\item{layout}{A vector indicating which layout to use as in par(mfrow = layout). If NA, will automatically generate an appropriate layout.}

\item{n_post}{An optional integer specifying how many posterior predictives to generate if the input is an emc object.}

\item{n_cores}{An integer specifying the number of cores with which to calculate predictives if specified.}

\item{...}{}
}
\value{
an invisible data frame with the stat applied to the real data, posterior predictives and/or prior predictives
}
\description{
Plots panels that contain a set of densities for each level of the specified \code{factor}
The densities represent the predicted data across the posterior, the vertical lines represent the real data.
}
\examples{
# For example plot the observed and predicted response accuracy
correct_fun <- function(data) mean(data$S == data$R)
plot_stat(samples_LNR, correct_fun, n_post = 10)
# Can also apply more sophisticated statistics
drt <- function(data) diff(tapply(data$rt,data[,c("E")],mean))
plot_stat(samples_LNR, drt, n_post = 10, stat_name = "RT diff Speed - A/N")

}
