% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{post_predict}
\alias{post_predict}
\title{Generate posterior predictives.}
\usage{
post_predict(
  samplers,
  hyper = FALSE,
  n_post = 100,
  filter = "sample",
  subfilter = 0,
  thin = 1,
  n_cores = 1,
  stat = c("random", "mean", "median")[1],
  ...
)
}
\arguments{
\item{samplers}{A list of samples from which you want to generate posterior predictives.}

\item{hyper}{Boolean. Default is FALSE, if true simulates from the group level instead of subject-level parameters. Only makes sense when aggregating across subjects}

\item{n_post}{Integer. How many data sets do you want to generate from the posterior.}

\item{filter}{Character. Choice of the stages 'preburn', 'burn', 'adapt', 'sample', default is 'sample'. From which stage do you want to take samples to generate new data with.}

\item{subfilter}{Integer or numeric vector. If integer, will filter out the first x of samples, within your filter. If numeric vector will select those samples, within your filter.}

\item{thin}{Integer. By how much do you want to thin the chains before simulating from them. Will keep 1/thin samples.}

\item{n_cores}{Integer. Across how many cores do you want to parallellize.}

\item{stat}{Character. Can be mean, median or default random. Will take either random samples from the chain or use the mean or median of the parameter estimates.}

\item{...}{Optional additional arguments}
}
\value{
A list of simulated data sets of length n_post.
}
\description{
Generates \code{n_post} data sets based on posterior parameter estimates
}
\examples{
\dontrun{
# based on a set of samplers ran by run_emc we can generate posterior predictives
post_predict(samplers, n_cores = 8)
}
}
