% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{profile_plot}
\alias{profile_plot}
\title{Creates a likelihood profile plot from a dadm object (see design_model) by
varying one model parameter while holding all others constant.}
\usage{
profile_plot(pname, p, p_min, p_max, dadm, n_point = 100, main = "", cores = 1)
}
\arguments{
\item{pname}{Name of parameter to profile}

\item{p}{Named vector of parameter values (typically created with sampled_p_vector)}

\item{p_min}{Minimum value of profile range}

\item{p_max}{Maximum value of profile range}

\item{dadm}{Data augmented design and model (created by design_model)}

\item{n_point}{Number of evenly spaced points at which to calculate likelihood}

\item{main}{Plot title}

\item{cores}{Number of likelihood points to calculate in parallel}
}
\value{
vector with value of p\link{pname}, highest likelihood point and p\link{pname}
minus the parameter values at that point
}
\description{
Creates a likelihood profile plot from a dadm object (see design_model) by
varying one model parameter while holding all others constant.
}
