% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{profile_plot}
\alias{profile_plot}
\title{Creates a likelihood profile plot from a design and the experimental data by
varying one model parameter while holding all others constant.}
\usage{
profile_plot(
  pname,
  p,
  p_min,
  p_max,
  data,
  design,
  n_point = 100,
  main = "",
  cores = 1
)
}
\arguments{
\item{pname}{Name of parameter to profile}

\item{p}{Named vector of parameter values (typically created with sampled_p_vector)}

\item{p_min}{Minimum value of profile range}

\item{p_max}{Maximum value of profile range}

\item{data}{A dataframe. Experimental data used, needed for the design mapping}

\item{design}{A design list. Created using \code{make_design}.}

\item{n_point}{Number of evenly spaced points at which to calculate likelihood}

\item{main}{Plot title}

\item{cores}{Number of likelihood points to calculate in parallel}
}
\value{
vector with value of p(pname), highest likelihood point and p(pname)
minus the parameter values at that point
}
\description{
Creates a likelihood profile plot from a design and the experimental data by
varying one model parameter while holding all others constant.
}
