% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_RDM.R
\name{RDMGBM}
\alias{RDMGBM}
\title{RDMGBM Model}
\usage{
RDMGBM()
}
\description{
Racing geometric Brownian first-passage model with start-point variability.
Equivalent parameterization to RDMSWTN without \code{sv}.
}
