% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_RDM.R
\name{RDMSWTN}
\alias{RDMSWTN}
\title{RDMSWTN Model}
\usage{
RDMSWTN(erlang_shape = 1L, erlang_type = "none", posdrift = TRUE)
}
\arguments{
\item{erlang_shape}{integer shape of the killing process (1=exponential, 2=Erlang-2) or "mixed"}

\item{erlang_type}{string, one of "none", "local_kill", "global_kill", "local_guess", "local_kill_guess"}
}
\value{
a list of parameters
}
\description{
RDMSWTN Model
}
\details{
Racing Diffusion Model with Shifted Wald Truncated Normal (SWTN) accumulators.
Supports between-trial drift variability (sv) and start-point variability (A).
When sv=0 and A=0 the model reduces to a point Wald; when sv=0 it reduces
to the standard RDM.\tabular{lllll}{
   \strong{Parameter} \tab \strong{Transform} \tab \strong{Natural scale} \tab \strong{Default} \tab \strong{Interpretation} \cr
   \emph{v} \tab log \tab [0, Inf] \tab log(1) \tab Mean drift rate \cr
   \emph{B} \tab log \tab [0, Inf] \tab log(1) \tab Response threshold (before A offset) \cr
   \emph{A} \tab log \tab [0, Inf] \tab log(0) \tab Start-point variability range \cr
   \emph{t0} \tab log \tab [0, Inf] \tab log(0) \tab Non-decision time \cr
   \emph{s} \tab log \tab [0, Inf] \tab log(1) \tab Within-trial SD of drift rate \cr
   \emph{sv} \tab log \tab [0, Inf] \tab log(0) \tab Between-trial SD of drift rate \cr
   \emph{mG} \tab log \tab [0, Inf] \tab log(1) \tab Local guessing timer mean \cr
   \emph{mK} \tab log \tab [0, Inf] \tab log(1) \tab Local killing timer mean \cr
   \emph{omega} \tab pnorm \tab [0, 1] \tab qnorm(.5) \tab Erlang-1 mixture weight (mixed only) \cr
}
}
