% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_RDM.R
\name{RDMSWTN}
\alias{RDMSWTN}
\title{The Racing Diffusion Model with Trial-Varying Drift (DSWTN)}
\usage{
RDMSWTN()
}
\description{
Model file to estimate a Racing Diffusion Model where each accumulator
follows a Wald distribution with trial-varying drift rates, modeled
by the DSWTN (Distribution Scalar Wald Truncated Normal) distribution.
}
\details{
This model replaces the standard RDM's single drift rate \code{v} and start-point
variability \code{A} with parameters for a distribution of drift rates:
\code{mu_drift} (mean of the drift distribution) and \code{sigma_drift_sq} (variance
of the drift distribution).

Default values are used for all parameters that are not explicitly listed in the \code{formula}
argument of \code{design()}.

#' Default values are used for all parameters that are not explicitly listed in the \code{formula}
argument of \code{design()}.They can also be accessed with \code{RDM()$p_types}.\tabular{llllll}{
   \strong{Parameter} \tab \strong{Transform} \tab \strong{Natural scale} \tab \strong{Default} \tab \strong{Mapping} \tab \strong{Interpretation} \cr
   \emph{v} \tab log \tab [0, Inf] \tab log(1) \tab  \tab Evidence-accumulation rate (drift rate) \cr
   \emph{A} \tab log \tab [0, Inf] \tab log(0) \tab  \tab Between-trial variation (range) in start point \cr
   \emph{B} \tab log \tab [0, Inf] \tab log(1) \tab \emph{b} = \emph{B} + \emph{A} \tab Distance from \emph{A} to \emph{b} (response threshold) \cr
   \emph{t0} \tab log \tab [0, Inf] \tab log(0) \tab  \tab Non-decision time \cr
   \emph{s} \tab log \tab [0, Inf] \tab log(1) \tab  \tab Within-trial standard deviation of drift rate \cr
   \emph{sv} \tab log \tab [0, Inf] \tab log(0.1) \tab  \tab Standard deviation of the trial-by-trial drift rate distribution \cr
}


All parameters are typically estimated on the log scale. \code{A} (start-point variability range)
and \code{sv} (drift rate between trial variance) can be zero.
\itemize{
\item If \code{A=0} and \code{sv=0}, the model reduces to a simple Wald distribution with threshold \code{B} and drift \code{v}.
\item If \code{sv=0} (and \code{A > 0}), the model reduces to the standard RDM (with start-point variability \code{A} and fixed drift \code{v}).
\item If \code{A=0} (and \code{sv > 0}), the model reduces to a SWTN with a fixed threshold \code{B}.
\code{B} should generally be positive, especially if \code{A=0}.
}

Like the standard RDM, this is a race model with one SWTN-spv accumulator per response option.
}
