% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{recovery_plot}
\alias{recovery_plot}
\title{Plots recovery of parameters}
\usage{
recovery_plot(
  tabs,
  layout = c(2, 3),
  do_ci = TRUE,
  ci_col = "grey",
  cap = 0.05,
  correlation = "pearson",
  cor_pos = "topleft",
  stat = "coverage",
  stat_pos = "bottomright",
  digits = 3
)
}
\arguments{
\item{tabs}{A list of matrices containing true and estimated parameters (with CIs), output from \code{plot_pars()}}

\item{layout}{A vector specifying the layout as in \code{par(mfrow = layout())}}

\item{do_ci}{A Boolean. Add CIs to the plot or not? Defaults to \code{TRUE}.}

\item{ci_col}{A character string. Color code of credible interval.}

\item{cap}{Numeric. Width of the credible interval caps (passed to \code{arrows()})}

\item{correlation}{String. \code{pearson} or \code{spearman}. Which correlation type should be computed
and included in the plot?}

\item{cor_pos}{Character string. Position of the correlation (passed to \code{legend()})}

\item{stat}{Character string. 95 percent \code{coverage} or \code{rmse}. Which statistic
should be included in the plot?}

\item{stat_pos}{Character string. Position of \code{stat} within the plot (passed to \code{legend()})}

\item{digits}{Integer. Number of digits included in the plot for the correlation and \code{stat}.}
}
\value{
Invisible list with RMSE, coverage, and Pearson and Spearman correlations.
}
\description{
Uses output from \code{plot_pars} to plot generating vs. estimated (median with credible
interval; CI) parameters to look at parameter recovery.
}
\examples{
# Make up some values that resemble posterior samples
# Normally this would be true values that were used to simulate the data
pmat <- matrix(rnorm(12, mean = c(-1, -.6, -.4, -1.5), sd = .01), ncol = 4, byrow = TRUE)
# The matrix needs to have the appropriate names
# Conventionally this would be created before one makes data with true values
colnames(pmat) <- c("m", "m_lMd", "s", "t0")
rownames(pmat) <- c("as1t", "bd6t", "bl1t")
# First use plot_pars
recovered_alpha <- plot_pars(samplers_LNR, selection = "alpha", pars = pmat, do_plot = FALSE)
# Now use the output to make the recovery plots
recovery_plot(recovered_alpha, correlation = "pearson", stat = "rmse")

# Similarly we can plot other group-level parameters
recovered_mu <- plot_pars(samplers_LNR, selection = "mu", pars = pmat[1,], do_plot = FALSE)
recovery_plot(recovered_mu, correlation = "spearman", stat = "coverage")
}
