% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LNR.R
\name{REXG}
\alias{REXG}
\title{The Ex-Gaussian Race Model}
\usage{
REXG()
}
\value{
A model list compatible with \code{design()}.
}
\description{
Model file for a race model where each accumulator follows an ex-Gaussian
distribution with free mean, Gaussian scale, and exponential tail.
}
