% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{run_emc}
\alias{run_emc}
\title{Generic function to run samplers with default settings.}
\usage{
run_emc(
  samplers,
  stage = NULL,
  iter = 1000,
  max_gd = 1.1,
  min_es = 0,
  min_unique = 600,
  preburn = 150,
  p_accept = 0.8,
  step_size = 100,
  verbose = FALSE,
  verboseProgress = FALSE,
  fileName = NULL,
  particles = NULL,
  particle_factor = 50,
  cores_per_chain = 1,
  cores_for_chains = length(samplers),
  max_trys = 50
)
}
\arguments{
\item{samplers}{A list of samplers, or a fileName of where the samplers are stored.}

\item{stage}{A string. Indicates which stage is to be run, either preburn, burn, adapt or sample. If unspecified will assume the next unrun stage.}

\item{iter}{An integer. Indicates how many iterations to run sampling stage.}

\item{max_gd}{A double. The maximum gelman diagnostic convergence allowed. Will stop burn-in if below this number.}

\item{min_es}{An integer. The minimal effective size required to stop sampling.}

\item{min_unique}{An integer. The minimal number of samples required. Only works in adaptation.}

\item{preburn}{An integer. Specifies how many iterations to run preburn stage.}

\item{p_accept}{A double. The target acceptance probability of the MCMC process. This will fine tune the width of the search space. Default = .8}

\item{step_size}{An integer. After each of these steps, the requirements will be checked if they are met and proposal distributions will be updated. Default = 100.}

\item{verbose}{Logical. Whether to print sampling related messages}

\item{fileName}{A string. If specified will autosave samplers at this location.}

\item{particles}{An integer. How many particles to use, default is NULL and particle_factor is used. If specified will override particle_factor}

\item{particle_factor}{An integer. Particle factor multiplied by the square root of the number of sampled parameters will determine the number of particles used.}

\item{cores_per_chain}{An integer. How many cores to use per chain. Parallelizes across participant calculations.}

\item{cores_for_chains}{An integer. How many cores to use across chains. Default is the number of chains.}

\item{max_trys}{An integer. How many times it will try to meet the finish conditions. Default is 50.}
}
\value{
A list of samplers
}
\description{
Calls \code{auto_burn}, \code{run_adapt} and \code{run_sample} in order with default settings.
}
