% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{run_samplers}
\alias{run_samplers}
\title{Generic function to run samplers for any stage and any requirements.}
\usage{
run_samplers(
  samplers,
  stage,
  stop_criteria,
  p_accept = 0.8,
  step_size = 100,
  verbose = FALSE,
  verboseProgress = FALSE,
  fileName = NULL,
  particles = NULL,
  particle_factor = 50,
  cores_per_chain = 1,
  cores_for_chains = length(samplers),
  max_trys = 20,
  n_blocks = 1
)
}
\arguments{
\item{samplers}{A list of samplers, could be in any stage, as long as they've been initialized with make_samplers}

\item{stage}{A string. Indicates which stage is to be run, either preburn, burn, adapt or sample}

\item{stop_criteria}{A list. Defines the stopping criteria and for which types of parameters these should hold. See run_emc.}

\item{p_accept}{A double. The target acceptance probability of the MCMC process. This will fine tune the width of the search space. Default = .8}

\item{step_size}{An integer. After each of these steps, the requirements will be checked if they are met and proposal distributions will be updated. Default = 100.}

\item{verbose}{Logical. Whether to print emc related messages}

\item{verboseProgress}{Logical. Whether to print sampling related messages}

\item{fileName}{A string. If specified will autosave samplers at this location.}

\item{particles}{An integer. How many particles to use, default is NULL and particle_factor is used. If specified will override particle_factor}

\item{particle_factor}{An integer. Particle factor multiplied by the square root of the number of sampled parameters will determine the number of particles used.}

\item{cores_per_chain}{An integer. How many cores to use per chain. Parallelizes across participant calculations.}

\item{cores_for_chains}{An integer. How many cores to use across chains. Default is the number of chains.}

\item{max_trys}{An integer. How many times it will try to meet the finish conditions. Default is 50.}

\item{n_blocks}{An integer. Will block the parameter chains such that they are updated in blocks. This can be helpful in extremely tough models with large number of parameters.}
}
\value{
A list of samplers
}
\description{
Used by \code{run_emc}, \code{auto_burn}, \code{run_adapt} and \code{run_sample}.
Will break if you skip a stage, the stages have to be run in order (preburn, burn, adapt, sample).
}
