% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{sampled_p_vector}
\alias{sampled_p_vector}
\title{sampled_p_vector()}
\usage{
sampled_p_vector(design, model = NULL, doMap = TRUE)
}
\arguments{
\item{design}{a list of the design made with make_design.}

\item{model}{a model list. Default is the model specified in the design list.}

\item{doMap}{logical. If TRUE will}
}
\description{
Makes an empty p_vector corresponding to model.
matchfun only needed in design if uses lM factor
}
