% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{savage_dickey}
\alias{savage_dickey}
\title{The savage dickey ratio.
Can be used to approximate the Bayes Factor for group level mean effects.
Note this is different to \code{compare_IC} since it only considers the group level mean effect and not the whole model.}
\usage{
savage_dickey(
  samplers,
  parameter = NULL,
  H0 = 0,
  filter = "sample",
  FUN = NULL
)
}
\arguments{
\item{samplers}{A list. A samplers object.}

\item{parameter}{A string. A parameter which you want to compare to H0. Will not be used if a FUN is specified.}

\item{H0}{An integer. The H0 value which you want to compare to}

\item{filter}{A string. Specifies which stage the samples are to be taken from "preburn", "burn", "adapt", or "sample"}

\item{FUN}{A character string. Specifies an operation to perform with multiple parameters. All operations must be space separated.
e.g. valid would be: "B - 0.5 * t0"}
}
\value{
The BayesFactor for the hypothesis against H0.
}
\description{
The savage dickey ratio.
Can be used to approximate the Bayes Factor for group level mean effects.
Note this is different to \code{compare_IC} since it only considers the group level mean effect and not the whole model.
}
