% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{savage_dickey}
\alias{savage_dickey}
\title{The savage dickey ratio.
Can be used to approximate the Bayes Factor for group level mean effects.
Note this is different to MD in \code{compare} since it only considers the group level
mean effect and not the whole model.}
\usage{
savage_dickey(
  samplers,
  parameter = NULL,
  H0 = 0,
  filter = "sample",
  subfilter = 0,
  fun = NULL,
  mapped = F,
  selection = "mu"
)
}
\arguments{
\item{samplers}{A list. A samplers object.}

\item{parameter}{A string. A parameter which you want to compare to H0. Will not be used if a FUN is specified.}

\item{H0}{An integer. The H0 value which you want to compare to}

\item{filter}{A string. Specifies which stage the samples are to be taken from "preburn", "burn", "adapt", or "sample"}

\item{subfilter}{An integer or vector. If integer it will exclude up until}

\item{fun}{A function. Specifies an operation to be performed on the sampled or mapped parameters.}

\item{mapped}{A boolean. Whether the BF should be calculated for parameters mapped back to the real design, only works with selection = 'mu'.}

\item{selection}{A string. Default is mu. Whether to do the operation on the alpha, mu, covariance, variance, or correlations.}
}
\value{
The BayesFactor for the hypothesis against H0.
}
\description{
The savage dickey ratio.
Can be used to approximate the Bayes Factor for group level mean effects.
Note this is different to MD in \code{compare} since it only considers the group level
mean effect and not the whole model.
}
