% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_SS.R
\name{SSRDEX}
\alias{SSRDEX}
\title{The hybrid Wald / ex-Gaussian race model of the stop signal task}
\usage{
SSRDEX(exgaussian_approximation = NULL)
}
\arguments{
\item{exgaussian_approximation}{Optional list passed to \code{\link[=stop_signal]{stop_signal()}}.
The default \code{NULL} disables the legacy ex-Gaussian normal and
shifted-exponential approximations.}
}
\value{
A model list with all the necessary functions to sample
}
\description{
Model file to estimate the hybrid Wald / ex-Gaussian race model of stop signal task data in EMC2.
}
\details{
Model files are almost exclusively used in \code{design()}.

Default values are used for all parameters that are not explicitly listed in the \code{formula}
argument of \code{design()}.They can also be accessed with \code{SSRDEX()$p_types}.\tabular{llllll}{
   \strong{Parameter} \tab \strong{Transform} \tab \strong{Natural scale} \tab \strong{Default} \tab \strong{Mapping} \tab \strong{Interpretation} \cr
   \emph{v} \tab log \tab [0, Inf] \tab log(1) \tab  \tab Evidence-accumulation rate (drift rate) of the go process \cr
   \emph{A} \tab log \tab [0, Inf] \tab log(0) \tab  \tab Between-trial variation (range) in start point of the go process \cr
   \emph{B} \tab log \tab [0, Inf] \tab log(1) \tab \emph{b} = \emph{B} + \emph{A} \tab Distance from \emph{A} to \emph{b} (response threshold) of the go process \cr
   \emph{t0} \tab log \tab [0, Inf] \tab log(0) \tab  \tab Non-decision time of the go process \cr
   \emph{s} \tab log \tab [0, Inf] \tab log(1) \tab  \tab Within-trial standard deviation of drift rate of the go process \cr
   \emph{muS} \tab log \tab [0, Inf] \tab log(.3) \tab  \tab Mean of Gaussian component of ex-Gaussian stop finish time distribution \cr
   \emph{sigmaS} \tab log \tab [0, Inf] \tab log(.025) \tab  \tab Standard deviation of Gaussian component of ex-Gaussian stop finish time distribution \cr
   \emph{tauS} \tab log \tab [0, Inf] \tab log(.05) \tab  \tab Mean (inverse rate) of exponential component of ex-Gaussian stop finish time distribution \cr
   \emph{tf} \tab probit \tab [0, 1] \tab qnorm(0) \tab  \tab Attentional lapse rate for stop process ("trigger failure") \cr
   \emph{gf} \tab probit \tab [0, 1] \tab qnorm(0) \tab  \tab Attentional lapse rate for go process ("go failure") \cr
   \emph{exgS_lb} \tab - \tab [-Inf, Inf] \tab 0 \tab  \tab Lower bound of ex-Gaussian stop finish time distribution \cr
}


All parameters are estimated on the log scale, with the exception of \code{tf} and \code{gf} which are estimated on the probit scale, and \code{exgS_lb} which is estimated on the natural scale.

The parameterization \emph{b} = \emph{B} + \emph{A} ensures that the response threshold \emph{b} is
always higher than the between trial variation in start point.

Conventionally, \code{s} is fixed to 1 to satisfy scaling constraints.

Because the ex-Gaussian stop signal model is a race model, it has one accumulator per response option.
EMC2 automatically constructs a factor representing the accumulators \code{lR} (i.e., the
latent response) with level names taken from the \code{R} column in the data.

For race models, the \code{design()} argument \code{matchfun} can be provided, a
function that takes the \code{lR} factor (defined in the augmented data (d)
in the following function) and returns a logical defining the correct response.
In the example below, the match is simply such that the \code{S} factor equals the
latent response factor: \code{matchfun=function(d)d$S==d$lR}. Then \code{matchfun} is
used to automatically create a latent match (\code{lM}) factor with
levels \code{FALSE} (i.e., the stimulus does not match the accumulator) and \code{TRUE}
(i.e., the stimulus does match the accumulator). This is added internally
and can also be used in model formula.

This hybrid race model of the stop signal task was introduced in Tanis et al. (2024).

Note that, in contrast to the parameters of the go process, the ex-Gaussian stop process parameters \code{muS}, \code{sigmaS}, and \code{tauS} do not have clear psychological interpretations (Matzke & Wagenmakers, 2009).
Matzke et al. (2020) showed that evidence accumulation models of the stop process have poor psychometric properties.
Thus, this hybrid race model - with an evidence accumulation account of \emph{going} and a descriptive account of \emph{stopping} - represents a compromise between process realism and practically useful measurement properties (Tanis et al., 2024).

The mean of the ex-Gaussian stop distribution (\code{muS + tauS}) is taken as the stop signal reaction time (SSRT).

The ex-Gaussian distribution has support on the real line \eqn{\left(-\infty, \infty\right)}.
To prevent evaluation of impossible (i.e., negative) or implausibly fast stop finish times, lower truncation is applied to the stop finish time distribution using the parameter \code{exgS_lb}.
The default value for this lower bound is \code{0}.
}
\references{
Matzke, D., & Wagenmakers, E. J. (2009). Psychological interpretation of the ex-Gaussian and shifted Wald parameters: A diffusion model analysis. \emph{Psychonomic Bulletin & Review}, \emph{16}, 798-817. \doi{10.3758/PBR.16.5.798}

Matzke, D., Logan, G. D., & Heathcote, A. (2020). A cautionary note on evidence-accumulation models of response inhibition in the stop-signal paradigm. \emph{Computational Brain & Behavior}, \emph{3}(3), 269-288. \doi{10.1007/s42113-020-00075-x}

Tanis, C. C., Heathcote, A., Zrubka, M., & Matzke, D. (2024). A hybrid approach to dynamic cognitive psychometrics: Dynamic cognitive psychometrics. \emph{Behavior Research Methods}, \emph{56}(6), 5647-5666. \doi{10.3758/s13428-023-02295-y}
}
